+++ src/components/SearchBar/Circles.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import styles from './Circle.module.css'
+
+function Circles() {
+  return (
+    <div className={styles.threeCircles}></div>
+  )
+}
+
+export default Circles
