+++ src/components/SideMenu/Nav.js
@@ -1,38 +1,32 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
+import NavIcon from './NavIcon'
 
 function Nav({ isLoggedIn, seteIfUpdateMySentencePage }) {
   const navStyle = {
     color: 'white',
   }
 
+  const linkClassName = `flex flex-row justify-between pt-3 pb-3 pr-2 pl-2 rounded hover:bg-slate-500 opacity-80`
+
   return (
     <React.Fragment>
       {/* <a href="/" style={navStyle}><span>Home</span></a> */}
-      <div className="side-home">
-        <Link to="/" style={navStyle}>
-          <svg version="1.0" xmlns="http://www.w3.org/2000/svg" width="92" height="40" viewBox="0 0 1280 544">
-            <path d="M4.9 28.6c-1 1.3-1 2 0 3.5.6 1.1 1 2.8.7 3.7-1 3.1 2.1 12.1 5.9 17.2C18 61.8 35.6 73.5 50 78.6c3.6 1.2 9.2 3.2 12.5 4.4 13.3 4.9 39.8 10.2 60 12.1 6.6.6 13.6 1.3 15.5 1.4 1.9.2 19.5.7 39 1 19.5.3 36.1.8 36.8 1.1 1.9.6-8.1 12-12.7 14.4-7.9 4.1-25.1 15.9-25.1 17.2 0 .4-1.6 1.4-3.5 2.2-1.9.8-3.5 2-3.5 2.5 0 .6-.6 1.1-1.4 1.1-.7 0-3.7 2.3-6.6 5.2-2.9 2.8-6.7 6.1-8.5 7.3-15.1 10.3-41.3 36.2-51.4 51-2.6 3.8-6.9 10.1-9.5 13.8-7.4 10.8-12.6 21.2-13.6 27.1-1.8 11.5 6.2 32.6 14.3 37.7 6.6 4.1 7.8 4.6 14.1 5.8 11.3 2.1 29-2.6 59.4-15.7l9.3-4-.6 3c-.4 1.7-.8 7-1 11.7-.3 10.2-1.2 14.7-3.5 18.1-.9 1.4-2 3.3-2.3 4.3-.6 1.5-7.8 6.6-8.5 5.9-.7-.7-6.2 3.1-6 4.1.2.7-1.3 2.7-3.2 4.5-1.9 1.7-2.9 3.2-2.2 3.2.6 0 1.2.8 1.2 1.8 0 2.4 1.9 4.2 3.6 3.5.8-.3 1.4.1 1.4 1.1 0 .9-.4 1.6-1 1.6-.5 0-1 .7-1 1.5s.7 1.5 1.5 1.5c.9 0 2 1.1 2.5 2.5 1 2.8 2.3 3.5 6.7 3.9 3.7.4 7.3 1.7 7.3 2.6 0 1.4 5 1.7 6.7.4 1.9-1.4 5.5-1.8 6.7-.7 1 1.1 8.9 1.3 10.1.3.6-.5 10.5-1.4 22-1.9 11.6-.6 30-1.5 41-2 11-.6 29.9-1.5 42-2.1 30-1.5 35.3-3.1 7.5-2.3-11.8.3-24 .7-27 .8-3 .2-16.5.5-30 .9-13.5.3-30.6.8-38 1.1-8.7.3-12.8.2-11.5-.4 1.1-.5 4.7-1 8-1.1 3.3-.1 7.8-.5 10-.9l4-.9-4.7-.1c-2.6 0-4.9-.5-5.2-1-1.3-2 9.2-3.2 28.2-3.1 10.8.1 21.1-.3 22.7-.7 3.3-1 15.8-.8 17.4.1.5.4 6.7.3 13.8 0 7-.4 14.8-.7 17.3-.8 10.1-.4 12-.7 12-1.7 0-.8-2.5-1-8.9-.5-6.2.5-9.1.4-9.6-.4-.4-.6-1.4-.8-2.4-.5-.9.4-5.6.8-10.4 1-4.8.2-9.7.7-10.8 1-1 .4-1.9.2-1.9-.3 0-.6-4.9-1.3-11-1.7-6-.4-11-1.1-11-1.6 0-1.3 14.8-1.1 15.2.2.2.7 1.2.7 3.1 0 1.7-.7 2.7-.7 2.7-.1 0 1.3 4.7 1.3 5.5 0 .5-.8 1.1-.8 2 0 1 .8 1.5.8 2 0 .4-.7 1.8-.9 3-.6 1.3.3 2.5.2 2.7-.3.2-.5 5-1.2 10.7-1.7 7.7-.5 10.2-.4 9.7.5-1.1 1.7 1.9 1.3 3.4-.4.7-.9 2.8-1.5 4.9-1.4 2.9 0 3.2.2 1.6.9-1.8.8-1.8.9-.2.9 1 .1 1.5.5 1.2 1.1-.3.6-.2 1 .3 1 .6 0 1.3-.7 1.6-1.6.4-1.1 1.3-1.4 3-1 1.4.3 2.8.1 3.1-.4.4-.6 1.2-.7 1.8-.3 2.1 1.3 4.8 1.1 4.3-.4-.4-.9.5-1.3 3.1-1.3 3 0 3.4.2 2.3 1.5s-.8 1.4 2.1.9c4.1-.8 5-1.4 2.1-1.4-1.2 0-2.2-.5-2.2-1 0-.6 2.6-1 5.9-1 3.3 0 6.6-.3 7.5-.6.9-.3 1.6-.1 1.6.5 0 .5-2.3 1.3-5 1.7-2.8.4-5 1-5 1.4 0 1 10.2 0 11.6-1.1.6-.5 1.9-.7 2.8-.3.9.3 1.6.2 1.6-.4 0-.6-.6-1.2-1.2-1.5-2.3-.8 12-2.8 31.7-4.3 4.4-.4 8.8-1.1 9.8-1.6 2.4-1.2 2.1 0-.5 2.3-2 1.6-2.1 1.8-.5 1.9 2.6 0 .2 3.4-4 5.4-4.9 2.3-4 3.7 1.2 1.9 5.3-1.7 8.5-1.2 6.5 1.2-.7.8-2.2 1.5-3.4 1.5-4.5 0-6.7 1.2-6.4 3.4.4 2.7 6.4 3.4 10.6 1.2 1.6-.8 3.8-1.5 4.8-1.5 2.5 0 5.5-.6 7.4-1.4.8-.3 2.4-.7 3.5-.9 1.3-.2 2.4-1.6 3.1-3.8.8-2.5 2-3.7 4.5-4.6 4.4-1.6 5.2-1.7 4.3-.3-.5.9.1 1 2.2.3 1.6-.4 4.7-.7 6.9-.6 5.9.3 10.5-.7 10.5-2.2 0-1.6-3.7-2.8-4.5-1.6-.3.5-1.8.2-3.3-.6-1.5-.7-3.5-1.4-4.4-1.4-1 0-1.8-.5-1.8-1.1 0-.8 1.9-1 6.5-.5 3.6.4 6.5.3 6.5-.2s1.3-.7 2.9-.4c2 .3 3-.1 3.5-1.4.8-2.2-.6-3.7-3-3.2-1.1.2-2.9-.8-4.2-2.2-2.3-2.4-2.3-2.4-1.9-.3.5 2 .3 2.2-1.6 1.1-1.2-.6-2.6-1.7-3.2-2.3-.5-.7-2.3-1.6-3.9-1.9-2.7-.7-6.5-4.1-6.6-6 0-.4 1.1-.5 2.5-.2 1.4.4 2.5.2 2.5-.3 0-.6.6-.9 1.3-.7.6.1.9.1.5-.1-.5-.3-.8-1.1-.8-1.9 0-.8-.7-1.7-1.5-2-2.5-1-1.7-2.4 1.4-2.4 2.9 0 5.5-1.9 2.9-2.1-.7-.1-2-.2-2.8-.3-1-.1-1.5-1.2-1.5-3.4 0-3.8-.1-3.9-2.5-.9-1.3 1.5-2.2 1.9-2.9 1.1-1.3-1.2-.4-7.4 1-7.4.5 0 .9-.7.9-1.5 0-2.2-1.6-1.8-3.4.7-2.3 3.2-3.2 1.3-1.3-2.7.9-1.9 1.4-3.5 1.1-3.5-1.6 0-5.7 3-12.5 9.2-4.2 3.7-8 6.8-8.5 6.8s-2.7 1.2-4.8 2.6c-2.2 1.5-5.3 3-7 3.4-1.7.4-4.2 1-5.6 1.4l-2.5.7 2.5-2.1 2.5-2.1-4 .6c-7.3 1.2-12 5.9-5 5.1 1.9-.3 3.7 0 4 .5.4.5 1.3.5 2.3 0 2.6-1.3 12.2-2.1 12.2-1 0 .9-.8 1.3-6.5 3.5-1.1.4-3.3.9-5 1-6.4.5-12.5 1.7-12.5 2.5 0 .5-.6.6-1.3.3-.7-.2-2.4 0-3.8.5-1.7.7 0 .8 6.1.5 5.7-.4 9.1-1.1 10.3-2.1 1.2-1.2 1.7-1.3 1.7-.3 0 .7.8 1 2 .7 1-.3 2.4.1 2.9.7.6.7 2.4 1.1 4.1.8 1.6-.2 3 0 3 .6 0 1.4-2.7 2.3-4.7 1.5-1.3-.5-1.5-.2-1 1.1.4 1 .1 1.9-.7 2.2-1.8.7-1.7 4.4.1 4.4.7 0 1.3-.5 1.3-1 0-.6.5-1 1.1-1 .8 0 .8.4 0 1.4-1.7 2-6.6 2.9-9 1.7-1.3-.7-2.2-.8-2.6-.1-.3.5-1.5 1-2.6 1s-1.7-.4-1.4-1c.3-.6.1-1-.5-1s-1.3.6-1.6 1.4c-.4 1.1-2.6 1.5-8.7 1.6-8.8.1-10.3.2-24.7 1.9-13.1 1.6-37.9 3.1-39.7 2.4-1.2-.5-1.2-.7-.1-1.4 2.4-1.5-1.4-1-4.2.6-3.1 1.8-7.4 2-6.4.3.5-.8-1-1-6-.5-4.2.4-6.5.3-6.1-.3.3-.6 1.9-1 3.5-1s3.2-.5 3.6-1.2c.5-.8-1.6-1-8.5-.5-8.8.5-9.2.5-8.6-1.4.8-2.3-1-2.5-2.3-.2-1 1.8-1.2 1.8-8.2 2.7-2.5.3-3.6.3-2.5 0s2.4-1 2.9-1.5c.6-.5 1.9-.9 3-.9s2.7-.6 3.5-1.4c.9-.7 3.9-1.7 6.6-2 3.8-.5 4.6-.4 3.5.4-1.7 1.3 1.2 1 11.9-1.3 3.5-.8 7.2-1.2 8.2-1 .9.3 2.7-.1 3.9-.7 1.7-.9 2.4-.9 3 0 .8 1.3 10.5-.1 13.8-1.9 1.4-.8 1.3-1-1-1.6-1.5-.4-3.3-.3-4 .1-.7.4-4 .8-7.5.8-8.6.1-6.2-1.7 2.8-2.1 3.9-.1 8.7-.7 10.7-1.3 2.4-.6 3.8-.6 4 0 .2.6.8 1 1.3 1 .6 0 .7-.5.4-1-.3-.6-.1-1 .6-1 .8 0 .5-.8-.7-1.9-1.1-1.1-2.3-1.5-2.9-1-1.3 1.3-2.7 1.1-1.9-.3.5-.8.2-.9-.9-.5-.9.3-4 .9-6.9 1.3-2.9.4-5.9 1.1-6.5 1.7-.9.6-1.8.6-2.7-.2-2.2-1.8-10.4-.4-11.7 2-.9 1.7-.6 1.9 2.6 1.9 1.9 0 3.5.3 3.5.7 0 .9 0 .9-16 1.7-6.3.3-14.5.8-18.2 1.1-3.9.3-6.9.1-7.3-.5-.4-.6 1.6-1 5.3-1 7 0 26.2-1.9 27.5-2.7 2.1-1.3-3.5-3.4-7.2-2.8-4.2.7-27.5 1.5-40.7 1.4-4.4 0-12.1.8-17 1.7-4.9.9-9.8 1.7-10.9 1.7-1.1-.1-4.2.3-7 .9-2.7.5-6.3 1.1-8 1.4-1.6.2-4.8 1.2-7 2.3-7.4 3.4-9.6 4.1-13.9 4.1-3.6 0-4.5-.4-5.5-2.6-1-2.3-.9-2.6.9-3.1 1.9-.5 1.9-.6.3-1.8-1-.8-1.6-1.5-1.3-1.7.3-.2 1.4-.8 2.5-1.4 1.7-1 2-2.3 2.3-9.2.3-10.9 2.9-18.7 8.5-25.9 4.5-5.6 11.5-12.2 16.3-15.3 14.5-9.3 28.4-20.2 28.4-22.3 0-.4-.8-.7-1.7-.7-1.7-.1-1.7-.2-.2-1.3 1.4-1 1.5-1.6.4-3-.7-.9-2.2-1.7-3.3-1.7-1.1 0-2.4-.5-2.7-1.1-.4-.6-1.4-.8-2.2-.5-.9.4-2.1.2-2.7-.3-1.6-1.2-8.3-2.3-8.8-1.4-.7 1.6-6.4 5-9.5 5.6-2.7.5-3.3.3-3.3-1.1 0-1.7-4.8 3-10.8 10.7-1.5 1.9-3.7 4.6-4.9 6.1-3.4 4.2-46 24.7-58.4 28.1-5.3 1.5-19.7 1.3-21.7-.3-4.7-4-6.7-17.9-3.7-26.9 1.7-5.1 2.7-7.1 8.8-17.7 3.9-7 18-24.2 25.4-31.2 15.2-14.4 35.6-30.7 50.7-40.5 30.8-19.9 75.9-41 96.6-45 12.6-2.5 19.5-3.2 22.5-2.3 1.7.4 4.2.6 5.8.2 1.5-.3 2.7-.2 2.7.3s4.1.9 9.1 1c8.7.1 9 .1 8.7-2-.2-1.7.2-2.2 1.9-2.2 1.2 0 3.1-.8 4.3-1.8 1.1-.9 2.9-1.7 4-1.7 2.4 0 2.6-1.8.5-4.7-1.4-2-1.4-2.3.2-3.2 2.2-1.2 2.5-.9-6.9-5.5-4.2-2-8.4-4.6-9.4-5.7-.9-1-2.5-1.9-3.6-1.9-1 0-1.8-.6-1.8-1.4 0-2.7-2.5-3.6-7.6-2.7-3.9.8-5.1.6-5.7-.5-.4-.8-1.7-1.4-2.8-1.4-1.3 0-2.3-.8-2.6-2.1-.5-1.7-1.1-1.9-3.7-1.3-3.3.7-7.9 3.4-5.8 3.4.7 0 1 .5.7 1-.4.6-1 .8-1.5.5-.4-.3-1.3.2-2 .9-.6.8-2.6 1.8-4.3 2.2-1.8.3-3.6 1-4.2 1.4-2.6 2-75.9.4-91-2-3.4-.5-11.9-1.2-19-1.5-11.9-.5-26.3-1.4-45.5-3.1-4.1-.4-9.2-.6-11.3-.5-2.1.1-5.7-.4-8-1s-6.4-1.3-9.2-1.6c-4.2-.4-12.3-1.9-30.5-5.9-8.4-1.8-26.9-8.4-35.5-12.6-10.3-5.1-14.5-8.1-18.2-13.4-2-2.8-2.7-3-8.1-3.2-4.3-.1-6.4.3-7.3 1.4zM434 288c0 .5-.7 1-1.6 1-.8 0-1.2-.5-.9-1 .3-.6 1-1 1.6-1 .5 0 .9.4.9 1zm-15.8 9.2c-.7.7-1.2.8-1.2.2 0-1.4 1.2-2.6 1.9-1.9.2.3-.1 1.1-.7 1.7zm-156.5 2.4c-.3.3-1.2.4-1.9.1-.8-.3-.5-.6.6-.6 1.1-.1 1.7.2 1.3.5zm-36 2.1c-.4.3-2.7.9-5.2 1.3s-4.2.4-3.8-.1c.7-.7 3.5-1.2 8.2-1.8.8 0 1.1.2.8.6zm185.9 2c-1 1-1.9-.5-1-1.8.6-1.1.8-1.1 1.2 0 .2.7.1 1.5-.2 1.8zm-150.1 2.7c-2.6 1.7-11.5 3.1-11.5 1.7 0-.6 1.7-1.3 3.8-1.6 2-.4 4.4-.8 5.2-1 3.7-.7 4.4-.5 2.5.9zm171.5.8c0 1.6-.3 1.9-1.1 1.1-.6-.6-.8-1.6-.5-2.2 1.1-1.7 1.6-1.3 1.6 1.1zm-166.5-.2c-.3.5-1.3 1-2.1 1s-1.2-.5-.9-1c.3-.6 1.3-1 2.1-1s1.2.4.9 1zm175 .9c.3.5.1 1.2-.5 1.6-.5.3-1-.1-1-.9 0-1.8.6-2.1 1.5-.7zM873.5 56c-.3.5.1 1 .9 1s1.8-.5 2.1-1c.3-.6-.1-1-.9-1s-1.8.4-2.1 1zM880.2 56.2c.5 1.5 2.8 2.3 2.8 1 0-.5-.7-1.3-1.6-1.6-1-.4-1.4-.1-1.2.6zM1237.3 57.4c-2.4 1.2-4.2 2.8-4.3 3.7 0 .9-.6 1.9-1.4 2.2-1.1.4-7.8.1-12.1-.6-1.7-.3-.4 6.7 1.8 9.8 1.2 1.6 2.5 5.2 2.9 8 2.5 16.5 6.1 74.8 5.4 86.3-.3 4.3-.1 6 .7 5.5.6-.4.8 0 .4.9-.3.9-.1 2 .5 2.6.8.8.7 1.6-.3 2.8-.9 1.2-1 1.5-.1 1 .6-.4 1.2-.3 1.2.3 0 .5-.5 1.3-1.1 1.7-.8.4-.9 2.8-.5 7.3 1.8 18 1.8 20.2.4 21.8-.9 1.1-.9 1.4-.1.9.9-.5 1.2.3 1.2 3.1-.1 2.1.3 26.7.8 54.8.9 48.4 1 66.8.4 69-.2.5-.3 5.5-.2 11 .2 9.5.3 9.9 1.7 7.2 2-3.6 2.7-3.4 1.8.6-.4 1.7-.8 4.1-.8 5.2-.1 1.1-.4 2.4-.7 3-.3.5-.7 3.2-.7 6-.2 4.9-.2 5 1.4 2.5 1.5-2.4 1.5-2.4.9 2-1.2 8.6-1.5 11.5-1.8 18.2-.2 6.3-.1 6.7 1.8 6.2 1.8-.4 1.8-.4.4.8-2.4 1.9-3.4 5.9-1.8 6.9 1.2.7 1.2 1.2.1 3-1.6 2.5-2.5 24.8-1.1 27 .6.9.9-.4.9-4.2 0-3.5.6-6.2 1.6-7.5 1.5-1.8 1.5-1.8.8.6-.4 1.4-.7 5.5-.7 9.2-.1 6.6-.8 8.5-2.1 5.2-.4-.9-.9-1.4-1.3-1-.3.3.2 1.4 1.1 2.5 1.6 1.7 1.6 2.2.3 4.8-.8 1.5-1.9 8.2-2.6 14.8-1.7 16.9-8.1 47.4-9.8 46.8-.8-.3-1.3.3-1.3 1.8 0 3.5-2.8 16.7-4.7 22.1-.9 2.6-1.5 5-1.2 5.2.3.3 2.8-.1 5.5-.8 5.2-1.4 10-4 12.1-6.6 1.1-1.3 1.2-1.3 1.3 0 0 .9-.4 2-1 2.6-.6.6-.2 1 1.3 1.1 1.2 0 1.7-.1 1-.4-.8-.3-.5-1.5 1-4 1.2-2.1 2.3-3.1 2.5-2.3.4 1 6.5-7.3 6.7-9 .1-1.3 4.5-7.3 4.5-6.2 0 .7.5 1 1.2.6s.8-.3.4.5c-.4.6-1.2.9-1.7.6-.5-.4-.9-.1-.9.4 0 2.2 3.1 1 4-1.5 1.3-3.4 1.3-3.9-.4-2.5-1 .9-1.6.9-2.1.1-.4-.6-.4-1.4-.1-1.7.3-.4.6 0 .6.8.1 1 .4.9 1.1-.4.6-.9 1.4-1.5 1.8-1.3 1.1.7 7.6-21.2 8.6-28.9.2-1.7 1.3-9.1 2.4-16.5 1.1-7.4 2.3-18.6 2.7-24.7.3-6.2 1-11.3 1.5-11.3s.9-.7.9-1.6c0-.8-.4-1.2-1-.9-.6.4-.9-1.5-.7-4.8.1-3 .5-5.3.8-5.1.4.2.9-1.1 1.3-2.9.4-1.9.2-3.3-.4-3.5-1.6-.6-1.2-2.7 1-5.5 1.1-1.4 2-3.1 2-3.9-.1-.7-.7-.2-1.5 1.2l-1.3 2.5-.8-2.9c-.4-1.7-.2-3.6.5-4.7.6-1 1.1-3.6 1-5.6-.1-3.4-.2-3.5-.9-1.3-.8 2.5-.8 2.5-1-.5-.5-6.4-1.8 7.3-1.9 20.9-.1 7.6-.3 14.1-.5 14.5-.2.3-.7 4.8-1 10-1 15.4-2.5 27-4.1 30.7-.8 1.9-1 3.2-.5 2.9.6-.3 1-.2 1 .4 0 .5-.6 1.2-1.2 1.4-1 .4-1 .6 0 .6.6.1 1.2.6 1.2 1.2 0 .5-.4.7-1 .4-.6-.4-.7.1-.3 1 .3 1 0 2.1-.8 2.6-1.2.7-1.2.9.1.9 1.1.1 1 .4-.4 1.5s-1.6 1.5-.5 1.5c.8 0 1 .3.4.8-.5.4-1.1 1.6-1.2 2.7-.6 5.7-2.3 11.5-3.3 11.5-.5 0-.6.8-.3 1.8.2 1.1-.9 4-2.6 6.7-3 4.7-4.3 5.5-2 1.4.6-1.1 1.4-4.1 1.9-6.7.5-2.6 1.1-6 1.5-7.6.5-2.6.4-2.8-.9-1.7-1.4 1.2-1.4.9-.5-2 .7-1.9 1.6-3.9 2.2-4.4.7-.7.7-.5.2.4-.5.9-.4 1.2.2.8.6-.4 1.3-1.6 1.6-2.7.4-1.7.1-2-1.6-1.8-1.8.2-2.2-.2-2-2.2.2-1.4.5-4.8.7-7.5.2-2.8.7-5.7.9-6.5.9-2.5 1.6-9 2.1-18 .3-5.6.2-7-.3-4.3-1 5.8-2.5 6.9-1.8 1.3 1.2-9 1.2-25.5 0-25.3-.8.2-.7-.6.2-2.4.9-1.7 1.4-6.3 1.4-13.3.1-5.9.4-15.7.7-21.9.4-8.2.2-11.1-.6-11.1s-1-1-.6-3.3c1-5 1.5-29 .6-28.1-.5.5-1.1 4-1.3 7.9-.3 3.8-1 8.1-1.6 9.5-.9 2.3-1 2.1-1-1.8-.1-2.3.2-4.2.7-4.2.4 0 .7-1.9.6-4.3-.2-6.4-.3-6.9-1.6-6.1-.7.4-.3-.3.9-1.4 2.9-2.8 4.3-8.2 4.3-16.1 0-5.3-.2-6.2-1.4-5.2-1.1.9-1.5.9-1.9-.1-.4-1-.8-1-1.6.2-.9 1.2-1.1 1.1-1.1-.7 0-1.5.5-2.1 1.4-1.7 2.8 1.1 4.6-2.5 4.6-9.2 0-8.8-1.8-8.1-2.2.8-.2 4.3-.7 7.2-1.3 6.9-.5-.1-.8-.8-.6-1.5.4-1.4 1-8.9 1.1-13.9 0-2.5.4-3.7 1.5-3.7 1 0 1.3-.9.9-3.6-.4-3.3-.2-3.5 1.2-2.3 1.3 1 1.7 1 2.4-.1.5-.8.7-1.7.4-2.1-.2-.4.1-1.1.7-1.5.8-.5.8-.9-.1-1.4-.9-.6-.9-1.3.1-3.1.7-1.3 1.2-3.5 1.2-4.9-.1-1.5 0-1.8.3-.8s1 1.8 1.5 1.8c.6 0 .8-.4.5-.9-.3-.4.4-2.5 1.4-4.7 1.1-2.1 2-4.4 2-5 0-.7.5-1.4 1.2-1.6.7-.3.9-1.1.3-2.4-1-2.2-1-2.1-.1-5.8.6-2.6.5-2.8-.8-1.7-1.3 1.1-1.4.8-.9-2.6.3-2.1 1-4.6 1.5-5.6.5-.9.6-2.2.3-2.8-.4-.5 0-2.1.9-3.4 2-3.1 2.1-5.7.1-3.1-.8 1.1-1.8 1.6-2.2 1.3-.3-.4.4-1.9 1.8-3.3 1.8-1.9 2.4-3.8 2.5-7.2.1-2.6-.3-5-.9-5.4-.7-.4-.7-1.1.2-2.1.6-.8.8-1.7.3-1.9-.6-.2-.7-2.4-.2-5.3 1-6.3.7-8-1.4-7.9-1 .1-1.6-.6-1.6-2 0-1.7.2-1.9 1.1-.6.7 1.1.8.1.3-3.5-.5-2.8-.9-8.4-1-12.5-.1-4.1-.6-17.6-1.2-30-.6-12.4-1.3-29.4-1.4-37.8-.1-8.3-.6-15.2-1-15.2-.3 0-.9-2.6-1.3-5.8-.4-3.1-1-7.9-1.5-10.7-.5-3.4-.4-7.1.6-11.8.8-3.7 1.2-7.2.9-7.7-.3-.6-.2-1.7.3-2.7 2.8-5-18.6-7.2-27.5-2.9zm27.7 197.1c0 .8-.4 1.5-1 1.5-.5 0-1-.7-1-1.5s.5-1.5 1-1.5c.6 0 1 .7 1 1.5zm2.3 0c-.3.9-.8 1.4-1 1.1-.3-.3-.2-.9.2-1.5.9-1.6 1.5-1.3.8.4zm-10.6 14.2c-.3 1-.5.4-.5-1.2 0-1.7.2-2.4.5-1.8.2.7.2 2.1 0 3zm-8.7 10.8c0 7-.4 11.5-1 11.5-.5 0-.8-1-.6-2.3.3-1.2.5-6.4.5-11.5.1-5 .3-9.2.6-9.2s.5 5.2.5 11.5zm7.6 3.5l-1.4 6.5-.1-6.7c-.1-3.7.4-7 .9-7.3.6-.4.8-1.1.5-1.6-.4-.5-.3-1.5.2-2.2 1.5-2.2 1.4 4.5-.1 11.3zm3.7-5.5c-.3.9-.8 1.4-1 1.1-.3-.3-.2-.9.2-1.5.9-1.6 1.5-1.3.8.4zm-.3 4.6c0 1.1-.4 1.8-1 1.4-.5-.3-.7-1.2-.3-2 .7-2.1 1.3-1.9 1.3.6zm-11.3 21.6c-.2 2.1-.4.4-.4-3.7 0-4.1.2-5.8.4-3.8.2 2.1.2 5.5 0 7.5zm4-2.5c-.3.8-.6.5-.6-.6-.1-1.1.2-1.7.5-1.3.3.3.4 1.2.1 1.9zm-8.8 14.2c.1 5.8-1.2 3.5-1.5-2.7-.2-3.6 0-4.6.6-3.2.5 1.1.9 3.8.9 5.9zm11.8 3.8c-.3.7-.5.2-.5-1.2s.2-1.9.5-1.3c.2.7.2 1.9 0 2.5zm-12 9c-.2 1.8-.4.6-.4-2.7 0-3.3.2-4.8.4-3.3s.2 4.2 0 6zm12.3-2.7c0 1.9-.4 3.5-1 3.5-.9 0-1.1-4.6-.3-6.3.8-1.5 1.3-.6 1.3 2.8zm-16 .5c0 1.1-.4 2-.9 2s-.7-.9-.4-2c.3-1.1.7-2 .9-2 .2 0 .4.9.4 2zm-.4 6.6c-.4.9-.9 1.5-1.1 1.2-.3-.3-.3-1.3.1-2.2.4-.9.9-1.5 1.1-1.2.3.3.3 1.3-.1 2.2zm13.7-1.1c-.3.9-.8 1.4-1 1.1-.3-.3-.2-.9.2-1.5.9-1.6 1.5-1.3.8.4zm-4.1 3c-.2 1.4-.6 2.5-.8 2.5-.2 0-.4-1.1-.4-2.5s.4-2.5.9-2.5c.4 0 .6 1.1.3 2.5zm-5.5.7c-.3.8-.6.5-.6-.6-.1-1.1.2-1.7.5-1.3.3.3.4 1.2.1 1.9zm-4 3c-.3.8-.6.5-.6-.6-.1-1.1.2-1.7.5-1.3.3.3.4 1.2.1 1.9zm0 9c-.3.7-.5.2-.5-1.2s.2-1.9.5-1.3c.2.7.2 1.9 0 2.5zm5 9.5c-.3 1-.5.2-.5-1.7s.2-2.7.5-1.8c.2 1 .2 2.6 0 3.5zm9.7 3.1c-.7 4.1-2.2 6.5-1.8 3 .1-1.3.3-3.6.3-5.1.1-1.7.5-2.6 1.1-2.2.6.4.8 2.2.4 4.3zm-9.7 13.5c-.2 2.3-.3.1-.3-4.8 0-5 .1-6.8.3-4.2.2 2.6.2 6.7 0 9zm10.2-1.1c0 .7-.4 2.2-.9 3.3-.7 1.7-.9 1.5-.9-1.3-.1-1.7.4-3.2.9-3.2.6 0 1 .6.9 1.2zm-8.3 9.2c-.8 5.2-2.6 7.2-2.6 2.9 0-2.8 2.1-8.5 2.9-7.8.2.2 0 2.4-.3 4.9zm8.5 13.5c-.6 8-1.6 14.8-2.1 14.9-.6.2-.8.7-.6 1.1.3.4.1 1.3-.5 2.1-.6 1.1-.9.4-.9-2.2 0-2 .5-4 1-4.3 1.2-.8 1.3-4.5.1-4.5-.5 0-.7-.9-.4-2 .3-1.1.9-2 1.4-2 .5 0 .9-.7.9-1.5s-.6-1.2-1.4-.9c-1.2.5-1.2-.4.1-5.6 1.4-5.3 1.4-6.3.2-7.1-1.2-.8-1.1-1.2.5-2.4 1.9-1.4 1.9-1.5.1-1.5-1.8 0-1.8-.1.1-1.4 1.7-1.3 1.9-1.3 2.2.7.2 1.1-.1 8.6-.7 16.6zm-17.2.5c-.9.8-1 .7-.5-.7.3-1 .7-3.1.8-4.5.1-2.2.2-2.1.5.7.3 1.9-.1 3.8-.8 4.5zm7.8 17.1c-.2 14.1-.7 24.9-1.3 25.5-.6.6-.9-1.1-.8-4.8.1-3.1.3-14.6.4-25.5.1-11.6.5-19.7 1.1-19.7.6 0 .8 9.6.6 24.5zm-5.7.5c0 1.6-.4 3-.9 3-.7 0-.8-1.9-.2-5.3.3-1.9 1.1-.4 1.1 2.3zm-3.3 10.2c-.2 1.3-.4.5-.4-1.7-.1-2.2.1-3.2.4-2.3.2 1 .2 2.8 0 4zm18-1.5c-.3 1-.5.4-.5-1.2 0-1.7.2-2.4.5-1.8.2.7.2 2.1 0 3zm-3.6 11.6c-1 .9-1.1.8-.5-.6.3-1 .9-1.5 1.2-1.2.3.3 0 1.1-.7 1.8zm-5.1 13.2c-.7.9-1.7 3.3-2.1 5.3-.6 3.3-.7 3.4-.8.9-.1-2.8 2.8-9.1 3.7-8.2.3.3-.1 1.2-.8 2zm4.6 3.7c-.2 3.5-.4 7.3-.5 8.5-.1 1.3-.6 2.3-1.1 2.3-.6 0-.8-1.1-.3-2.8.4-1.5.6-3.6.6-4.7-.2-2.7.8-9.5 1.3-9.5.2 0 .2 2.8 0 6.2zm-1.9 15c-.3.7-.5.2-.5-1.2s.2-1.9.5-1.3c.2.7.2 1.9 0 2.5zm-6.7 7.2c0 .3-.4.8-1 1.1-.5.3-1 .1-1-.4 0-.6.5-1.1 1-1.1.6 0 1 .2 1 .4zm9.4 7.1c-.4.8-.8 1.5-1 1.5-.2 0-.4-.7-.4-1.5s.4-1.5 1-1.5c.5 0 .7.7.4 1.5zm-2.4 22.9c0 .3-.4.8-1 1.1-.5.3-1 .1-1-.4 0-.6.5-1.1 1-1.1.6 0 1 .2 1 .4zm-2.6 6.5c-.4 1.7-.9 3-1.1 2.7-.3-.2-.2-1.9.2-3.6.9-4.1 1.7-3.3.9.9zM870 57.1c0 .5.5.7 1 .4.6-.3 1-.8 1-1.1 0-.2-.4-.4-1-.4-.5 0-1 .5-1 1.1z" />
-            <path d="M895.6 57.9c-.3.4-1.9 1.1-3.6 1.4-2.3.5-2.8 1-2.3 2.4.6 1.5.4 1.6-.9.5-1.7-1.3-4.8-.7-4.8 1.1 0 .8-.3.8-1.1 0-.9-.9-1.5-.8-2.6.3-1.9 1.9-3.1 1.8-3.6-.4-.3-1-.6-.5-.7 1.3-.2 2.9-.2 2.9-1.4-.9-2-6.3-2.6-6-7.1 3.4-.9 1.9-1.7 5.2-1.8 7.2-.1 2.1-.2 3.8-.4 3.8-.1 0-1.1-1.7-2.2-3.8-1.1-2-2-3-2.1-2 0 .9.5 2 1 2.3.6.3 1 2 1 3.6 0 1.8-.4 2.8-1 2.4-.5-.3-1 .1-1 .9 0 2.4 1.8 2.8 2.5.6.5-1.6 2.5-2.9 2.5-1.6 0 1.3-3.1 6-5.4 8.1-1.5 1.4-2.4 3-2.1 3.5.8 1.3 2.5.3 2.5-1.6 0-.8.5-1.4 1.1-1.4.7 0 .9 1 .4 2.7-.4 1.6-.7 4.3-.7 6.2.1 3.2-5.6 14.6-9.3 19.1-.7.8-3.2 5-5.4 9.3-5 9.6-11.7 18.7-13.6 18.7-.8 0-1.5.9-1.5 2s-.4 2-1 2c-1.4 0-1.2-3 .4-6.1 1-2 1.1-4.3.5-9.5-.4-3.8-1.3-11.4-1.9-16.8-.6-5.7-2.3-13.8-4-18.8-1.7-4.9-3-9.9-3-11 0-2.7-6.3-10.4-10-12.3-1.6-.8-4-1.5-5.4-1.5-3.4 0-4.2-1.8-.9-2.2 2.2-.2 2.4-.5 1.2-1.5s-2.1-.9-4.4.3c-3.6 1.8-6 .2-2.9-2.1 1.4-1.1 1.5-1.4.4-1.5-1.9 0-8 3.7-8 4.9 0 .5-1 1.1-2.2 1.3-1.3.2-2.2.9-2 1.6.1.7-.3 2.2-.9 3.4-.9 1.6-1.3 1.8-2 .7-.6-1.1-.9-1-1.5.4-.3.9-1.2 1.7-1.9 1.7s-3 2-5.1 4.5c-2.2 2.5-10 10.7-17.3 18.2C743.1 120 721 145 721 147c0 .6-.5 1-1.1 1-.7 0-.9-1.1-.6-2.8.3-1.5.1-6.3-.4-10.7-1.2-9.7-6-20.5-12.1-27.1-4.5-4.9-15.4-10.6-19.1-10-2.5.3-7.9-3.2-7-4.7.3-.5 0-.7-.8-.4-.8.3-3.7 1.3-6.6 2.2-2.8.9-5.3 2.2-5.5 3-.3 1.3-17.5 15.6-27 22.4-2.6 1.9-5 4.2-5.3 5.3-.4 1-1 1.8-1.5 1.8-1.1 0-8 7-16 16.4-3.6 4.2-7.1 7.6-7.7 7.6-.7 0-1.3.4-1.3 1 0 1.9-1.9 1-2.5-1.3-.9-3.2-2.7-6.7-3.5-6.7-1.5 0-8.5-7.4-8.2-8.6.2-.8-1.4-1.5-4.4-2-3.8-.5-4.5-1-4.1-2.5.4-1.4 0-1.9-1.3-1.9-1.1 0-2.2.4-2.5 1-.3.5-.1 1 .4 1 .6 0 1.1.4 1.1.8 0 .5-3.3 1.4-7.2 2.1-4 .7-8.7 1.6-10.4 2.2-3.9 1.2-5.7 1.2-4.9-.1.3-.6 1.5-1 2.6-1 1 0 1.9-.5 1.9-1 0-1.4-5.2-.1-6.6 1.6-.6.8-1.6 1.4-2.1 1.4s-1.8.9-2.8 2c-1 1.1-2.5 2-3.4 2-.9 0-2.5.7-3.5 1.5-1.1.8-2.6 1.5-3.4 1.5-.8 0-7 2.7-13.9 6s-13 5.7-13.6 5.3c-.5-.3-.7-.2-.4.4.6 1-6.5 5.3-8.9 5.3-.7 0-1.8.9-2.4 2-.6 1.1-1.7 1.9-2.4 1.8-1.2-.3-13.3 8.5-18.9 13.7-3.6 3.2-9.7 11.1-9.7 12.4 0 .6.7 1.4 1.6 1.8.8.3 1.3 1.2 1 2-.6 1.5 2.7 5.3 4.6 5.3 1.8.1 4.9 2.6 4.1 3.3-.3.4.1.7 1 .7s1.9-.5 2.2-1c.3-.6 1.6-1 2.7-1 3.4 0 9.2-2.3 10.6-4.3.8-1.1 2.1-1.6 3.2-1.2 1.3.4 1.7.1 1.6-1.2-.1-1.6 8.1-7.6 23.1-16.7 7.6-4.7 25.9-13 36.3-16.5 11.1-3.8 23.4-5.1 26.3-2.9 3 2.2 3.3 9 .7 14-2.9 5.7-15.7 23.5-18 25.1-1.2.8-1.6 1.8-1.1 2.6 1.3 2 0 6.1-2.6 8.7-1.7 1.7-2.3 3.2-2 4.6.3 1.1-.1 3.1-.9 4.3-1.2 1.9-1.2 2.6.1 5.8.8 2 1.5 5.7 1.5 8.2 0 4.6 1.2 5.8 2.8 2.9.8-1.5 1-1.5 1.5 0 .7 1.8 2.7 2.2 2.7.5 0-1.6 4-6.9 5.1-6.9.6 0 .7.4.4 1-.3.5-.2 1 .3 1 1.1 0 2.5-2.6 2.7-5.1.1-.9 2.8-5.6 6-10.5s7.9-13 10.3-17.9c2.5-5 5.1-9.7 5.7-10.5 1.4-1.7 3.5-9.8 3.5-13.3 0-1.4 1.7-4.3 4-7 4.3-4.8 5.4-7.2 1.9-4.1-4.7 4.1-1-.9 7.1-9.7 8.6-9.4 20.5-19.9 22.4-19.9.6.1-.2 1.2-1.9 2.5-5.7 4.7-1.8 3.4 6.4-2.2 8.7-5.9 18.6-9.2 27.6-9.3 7.8 0 9.7 3.5 11.1 20.8.4 4.9 1.2 8.3 1.9 8.6 2.2.8.4 27.1-2.7 39.1-.5 2.2-1.8 4.9-2.8 6-1.2 1.3-1.6 2.5-1.1 3.3.9 1.4-1.3 8.7-3.2 11-.8 1-.7 1.6.3 2.2 1.7 1 .6 2.5-1.7 2.5-.9 0-1.8.6-2.1 1.4-.3 1 0 1.2 1.1.8 1.2-.4 1.5 0 1.3 2-.2 1.3-.9 2.9-1.6 3.3-1.1.7-.9 1.3.9 2.7 2.1 1.7 2.1 2 .7 3.2-1.4 1.2-1.2 1.7 1.2 4.7 3.3 4.1 5.2 4.4 5.2.6v-2.7l1.9 2.4c1 1.3 1.6 3.2 1.4 4.1-.3 1 0 2.1.6 2.4.6.4 1.1.1 1.1-.6 0-.8.5-1.9 1.2-2.6.6-.6 1.3-2 1.5-2.9 1.1-7.2 1.6-8.3 3.8-8.9 1.7-.4 2.8-2.1 4.4-6.3 1.2-3.2 2.1-6.4 2.1-7.1 0-1.4 1.6-6.8 3.1-10.5.4-1.1 1.8-4.9 3-8.5 2.5-7.4 6.9-16.7 10.8-22.5 1.5-2.2 4.4-7.8 6.5-12.5 5.4-12.2 15.9-30 21.1-35.7 2.5-2.6 4.5-5.3 4.5-6 0-.7.9-1.7 2-2.3 1.1-.6 2.3-2.2 2.6-3.6.4-1.4.9-2.2 1.2-1.9.9.8 4.2-2.7 4.2-4.5 0-2.2 1.9-4 3.6-3.3.9.3 1.4-.2 1.4-1.7 0-1.6.4-2 1.5-1.6.9.3 1.5.1 1.4-.7-.2-2 1.3-5.7 2.1-5.2.5.3 1.1.1 1.5-.5.3-.5.1-1-.5-1s-.8-.5-.5-1c.5-.8 1.1-.7 2.1.1 1.1.9 1.4.7 1.4-.9 0-1.2.5-2.4 1.1-2.8.8-.4.9.1.4 1.6l-.7 2.3 3.5-2.1c4.2-2.4 4.9-1.9 6.6 4.5.7 2.5 1.8 4.3 2.6 4.3 2.3 0 8.1 19.8 10.1 34.5 1 7.1 1.9 19.4 2 27.3.2 12.4 0 14.3-1.3 13.8-2.7-1.1-3.7.6-1.3 2.4 2.7 2 2.1 5-.9 5-2.5 0-3.7 1.8-1.7 2.6.9.3 1.6 1 1.6 1.6 0 .5-.5.6-1 .3-1.2-.7-4 1.2-4 2.7 0 .6.7.8 1.5.5.9-.4 2.2.5 3.5 2.2 2.4 3.4 2.6 4.9.5 3.1-2.6-2.2-4.3.4-2 3 1 1.1 2.4 2 3.1 2 1.7 0 2.2 1.9 1.6 5.7-.3 2.1 0 3.3.7 3.3.6 0 1.1-.6 1.2-1.3 0-.6.9.7 1.9 3 1.1 2.4 2.3 4.3 2.8 4.3 1.5 0 6.2-6.2 5.5-7.3-.3-.5-.8-.4-1.3.3-.8 1.3-1.7-.4-3-5.8-.6-2.3-.5-3.2.3-2.9.7.2 1.3 1.4 1.5 2.6.3 1.7.6 2 1.7 1.1 1.1-.9 1.5-.2 2.1 3.5.6 4 .4 4.8-1.5 6.1-1.3 1-2.1 2.5-2 4.2.1 2.6.1 2.6.7.4.4-1.3 1-2.3 1.5-2.3.4.1 1.2.1 1.7.1.6 0 1 1.2 1 2.6 0 4.1 1.8 1.7 2.6-3.4.3-2.6.9-5.2 1.3-5.7.3-.6.7-3.4.7-6.2.2-7.2 1.3-11.1 2.5-9.2.7 1.1.9 1.1.9-.4 0-1 .4-1.5 1-1.2.5.3.8 2.5.6 4.8-.4 5.1.8 7.9 3.1 7 1.3-.5 1.5-.1 1.1 2-.3 1.5-.1 2.7.3 2.7.5 0 .9-.9.9-2 0-1.2.7-3.2 1.6-4.5 1.2-1.8 1.3-2.5.4-2.5s-.9-.4-.2-1.3c.6-.7.9-5.2.8-10.7-.2-8.9 0-9.9 3-16.2 3-6.1 5.5-9.7 10.9-15.3 1.7-1.8 4.1-4.7 13.5-16.2.5-.6 2.8-4.2 5.1-8 5.1-8.3 7.2-11.2 8.1-11.3.3 0 1 3.5 1.4 7.7.5 6.4.8 7.3 1.6 5.3.5-1.4.7-3.6.4-5-.2-1.4-.1-3.2.5-4 .7-1.3.9-1.3 1 0 .1.8.2 1.9.4 2.5.1.5.3 2 .3 3.2.3 3.1 1.6 2.9 3-.5 1-2.3 1.1-1.9.7 3.3-.4 4-.1 6.8.9 8.4 1.1 1.9 1.1 3.1.3 4.9-.9 2-.8 2.6.5 3.6 1.2.9 1.5 2.3 1.2 5.1-.8 7.6.7 22.7 2.2 23.2 1.3.4 2.4-2.8 2.4-6.7 0-1.5.5-3.1 1.1-3.7.6-.6 1.2-4.9 1.5-9.6.2-4.7 1.1-10.2 1.8-12.4 1.7-4.5 2.1-1.6.5 4-.8 3-.7 4.2.6 5.9.8 1.2 1.5 1.8 1.6 1.3.3-5.8 1.1-9 2.4-9.7.8-.5 1.6-2.1 1.6-3.6.2-2.4.2-2.4.6-.4.2 1.2.9 2 1.4 1.7 1.2-.8 1.2-5.2 0-6-.6-.3-.8-2-.6-3.8.3-1.7.8-7.5 1.2-12.7.4-5.2.9-8.7 1.1-7.8.9 4.3 3.2.9 3.2-4.8 0-1.9.5-3.4 1-3.4.6 0 1-.9 1-2s-.4-2-1-2c-.5 0-1 .8-1 1.7 0 1.5-.2 1.6-1 .3-.7-1.1-1-.4-1 2.7 0 2.3-.5 4.5-1.1 4.9-.8.4-.9-.3-.5-2.3 1-4.5.9-20.1-.2-26-.7-4.4-.6-5.8.8-8.1 1.7-2.7 1.2-7.2-.7-7.2-.4 0-.8-1.3-.9-3-.1-4.1-2.1-6.1-5-5.4-1.5.4-2.4.2-2.4-.5 0-1.3-2.6-1.5-3.4-.2zm-133.4 50.3c-.7.7-1.2.8-1.2.2 0-1.4 1.2-2.6 1.9-1.9.2.3-.1 1.1-.7 1.7zM588 136c0 .5-.4 1-1 1-.5 0-1-.5-1-1 0-.6.5-1 1-1 .6 0 1 .4 1 1zm310.6 3.7c-1.1 1.1-1.9-.6-1.1-2.4.8-1.7.8-1.7 1.2 0 .3 1 .2 2.1-.1 2.4zm-69.6 4.7c0 1.9-.8 4.4-1.9 5.7l-1.8 2.3.5-3.2c.2-1.8.8-3.5 1.3-3.8.5-.3.9-1.4.9-2.5 0-1 .2-1.9.5-1.9s.5 1.5.5 3.4zm-16.1 6.3c-.1 6.3-1.2 10.2-2.3 8.5-1-1.7.3-13.2 1.5-13.2.5 0 .9 2.1.8 4.7zm-.8 18c-.9 1.6-1 1.4-1.1-1.2 0-2.1.3-2.6 1.1-1.8s.8 1.6 0 3z" />
-            <path d="M867 59.9c0 1.2.5 2.1 1 2.1.6 0 1-.6 1-1.4 0-.8-.4-1.8-1-2.1-.5-.3-1 .3-1 1.4zM875 59c0 .5.9 1 1.9 1 1.1 0 2.8.3 3.8.6 1.2.5 1.5.4.8-.3-1.5-1.5-6.5-2.5-6.5-1.3zM1142.7 63.4c-.8 2-4.7 2.7-5.4 1-.7-1.8-2.3-1.8-2.3 0 0 .9-.8 1.2-2.5.8-1.5-.2-2.5 0-2.5.7 0 .6.6 1.1 1.3 1.1 2.2 0 4.9 3.7 5.5 7.2.3 1.8.5 9.8.3 17.8-.2 12.9-1.8 25.5-5 40-.8 3.5-.8 3.5-1.8 1-.6-1.4-.8-3.1-.6-3.7.3-.7-.7-1.7-2.3-2.3-2.1-.8-2.4-1.4-1.6-2.2 1.7-1.7 1.5-4.5-.3-5.2-.9-.3-1.2-1.2-.8-2.2.5-1.3.3-1.6-1-1.1-1.1.4-1.7.1-1.7-.9 0-.8-1.1-1.9-2.5-2.4s-2.5-1.6-2.5-2.4c0-.8-.6-2.1-1.4-2.8-.7-.8-1.7-2.5-2.1-3.8-.4-1.4-1.8-3.9-3.1-5.6-1.3-1.7-2.4-3.8-2.4-4.6 0-2.3-1-2.8-3-1.6-1.4.9-2.2.9-3.4-.2-2.6-2.1-3-.4-.7 3.3 2.5 4.1 2.6 4.7.6 4.7-.8 0-1.5-.5-1.5-1.1 0-1.5-3.6-6-4.3-5.3-.3.3.1 1.7.9 3.2.8 1.5 1.3 3.6 1.3 4.7-.1 1.3-.5.7-1.3-1.5-.6-1.9-1.5-4.1-1.9-4.8-.5-.8-.4-1.9.4-2.8 1-1.2.9-1.6-.9-2.1-1.2-.3-2.2-.9-2.2-1.4 0-.5-.7-.9-1.5-.9-1.9 0-1.9-.1.6 5.7 4 9.4 5.2 13.5 3.9 14.3-.8.5-.9 1-.2 1.4.6.3 1.7 1.9 2.5 3.5 1.2 2.5 1.2 3.1.1 4.3-1.1 1.1-1.1 1.8 0 4 .8 1.5 1.7 4.2 2.1 6 .4 1.8 1.2 3.8 1.7 4.5.5.7.7 2.1.3 3.3-.5 1.5-.2 2 1.1 2 1.4 0 1.5.3.6 1.2-1.6 1.6-1.5 4.5 0 5 .9.3.8 1.1-.4 3-1.4 2.1-1.6 4.1-1.1 10.9.3 4.6.5 10.2.4 12.4 0 2.2.5 5 1.1 6.3 1 1.8.9 2.5-.1 3.2-.8.4-1 1.3-.7 1.9 1.1 1.7 2.6 1.3 2.6-.6 0-2.8 1.5-.3 1.6 2.8.2 3.4.1 3.4-3.2 1.3-1.3-.9-2.7-1.4-3.1-1.1-.8.8 1.6 3.7 3.1 3.7.6 0 2 .7 3.1 1.6 1.8 1.4 1.8 1.6.4 4.7-4.5 9-4.4 8.8-2.9 9.7 1.1.7.9 1.3-1.4 3.1-1.9 1.5-2.4 2.5-1.7 3.2 1.1 1.1 9.1-2.4 11.2-5 .6-.7 1.7-1.3 2.4-1.3s1.8-.6 2.5-1.4c.6-.8 1.8-1.2 2.5-1 .7.3 2.2-.3 3.3-1.4 1-1 3.1-2.2 4.5-2.6 1.4-.3 3.9-2 5.6-3.7 2.6-2.6 3.4-2.9 4.6-1.9.9.7 1.9.9 2.3.5.4-.4.1-1-.5-1.2-2.4-.9-1.3-1.9 2-1.9 2.8.1 3.2-.2 3-2.2-.2-1.7.2-2.3 1.5-2 1.7.2 8.2-5.5 8.2-7.3 0-.5.5-.9 1.1-.9.8 0 .9 1.1.5 3.1-.5 2.7-.4 3 .9 1.9s1.8-1 2.9.5c1.3 1.8 1.4 1.8 2.4-.1.7-1.4.6-2.6-.4-4.4-.7-1.4-1.3-3.1-1.4-3.8 0-.8-.9-2.5-2-3.9-2.3-2.9-2.6-6.8-.8-11 .6-1.5 1.2-3.2 1.2-3.6.1-.5 1.4-2.2 3.1-3.9 2.7-2.7 2.8-3.1 1.3-4.2-1.5-1.1-1.5-1.8-.1-8.1.8-3.9 1.9-7.7 2.3-8.6 1.8-3.4 3.3-22.7 2.7-33.7-.6-10.9-.6-11.3 1.3-10.7 1.6.5 2 .2 2-1.4 0-1.2-.6-2.1-1.2-2.1-1.4 0-9.8-10.2-11-13.3-.8-1.9-8.6-5.7-13-6.3-1.5-.2-2.8.2-3.1 1zM1107.5 97c.3.5.1 1-.4 1-.6 0-1.1.5-1.1 1.1 0 .6.6.9 1.4.6 1.4-.5 3.1 1.8 2.7 3.6-.1.8 3.5 6.5 4.9 7.7 1.1.9 5 8.1 5 9.1 0 1.6-6.5-6.8-11-14.3-3.9-6.5-4.9-9.8-3.1-9.8.6 0 1.3.4 1.6 1zm-4.5 4.9c0 .5 2.5 4.5 5.6 9.1 3.1 4.5 5.4 8.4 5.1 8.6-.5.5-8.3-10.5-11-15.4-.9-1.7-1.3-3.2-.7-3.2.5 0 1 .4 1 .9zm-2 5.6c0 .5-.5.3-1-.5s-1-2-1-2.5c0-.6.5-.3 1 .5s1 1.9 1 2.5zm-1 11.5c0 .5-.5 1-1.1 1-.5 0-.7-.5-.4-1 .3-.6.8-1 1.1-1 .2 0 .4.4.4 1zm22.1 8.3c-.8.8-1.1.5-1.1-.9 0-2.7 1.4-4.1 1.8-1.9.2.9-.1 2.2-.7 2.8zm-11.8 26c-.3.5.7 1.9 2.2 3.1 2.4 1.9 2.7 2.6 2 5.7l-.8 3.4-2.2-5c-2.4-5.6-3.9-12.8-2.7-13.9.8-.9 2.3 5.5 1.5 6.7zM862.6 66.6c-.3.9-.2 2.3.4 3.2.7 1.4.9 1 1-1.6 0-3.5-.5-4.1-1.4-1.6zM1023.8 69c-1 .9-1.8 2.2-1.8 2.8 0 .7-.4 1.2-.9 1.2-1.6 0-1.9 6.4-.5 10.3.7 2 1.1 4.4.7 5.3-.7 2 1.6 23.6 2.8 25 .5.6.9 2.2.9 3.5 0 1.3.3 4.2.6 6.4.2 2.2.6 7.8.9 12.5.2 4.7.6 10.7.9 13.5.3 2.7.7 8.8.7 13.4.1 4.6.8 9.5 1.5 11 1.3 2.7 2.2 19.6 1 20.8-1 1-.7 4.3.5 4.3.5 0 .7.5.3 1.2-.4.6.1 1.7 1.1 2.4s1.5 2 1.2 2.8c-.4.9-.2 1.8.3 2.1.5.3 1.4 2.9 2 5.9 1.1 5.6 3.3 9.6 4.6 8.4.7-.8 1.7-7.1 1.8-12 .1-2.6.2-2.6 1.9-.8.9 1.1 1.7 2.4 1.7 3 0 .5.4 1 1 1 .9 0 1.4-1.4 2.4-7.5.2-1.7 1.2-6.8 2.1-11.4.8-4.7 1.4-9 1.1-9.7-.3-.7 0-1.1.7-.8 1.7.6 1.2-32.2-.8-49.1-1.6-13.7-4.1-29.8-5.4-35.1-1-3.9-.1-5.7 2.5-4.8 2 .7-1.3-3.1-18.8-21.7-5-5.3-5.4-5.5-7-3.9zM1094.1 76.6c0 1.1.3 1.4.6.6.3-.7.2-1.6-.1-1.9-.3-.4-.6.2-.5 1.3z" />
-            <path d="M1097.9 84.7c-1.3 1.5-1.2 1.5.9.4 2.3-1.2 2.8-2.1 1.3-2.1-.5 0-1.5.7-2.2 1.7zM1086.5 86c.3.5 1.1 1 1.6 1 .6 0 .7-.5.4-1-.3-.6-1.1-1-1.6-1-.6 0-.7.4-.4 1zM1103 90.6c0 .8.5 1.2 1 .9.6-.3 1-1 1-1.6 0-.5-.4-.9-1-.9-.5 0-1 .7-1 1.6zM910 91.4c0 .9.5 1.6 1 1.6.6 0 1-.4 1-.9 0-.6-.4-1.3-1-1.6-.5-.3-1 .1-1 .9zM910 96c0 .5.5 1 1.1 1 .5 0 .7-.5.4-1-.3-.6-.8-1-1.1-1-.2 0-.4.4-.4 1zM765 105.1c0 .5.5.7 1 .4.6-.3 1-.8 1-1.1 0-.2-.4-.4-1-.4-.5 0-1 .5-1 1.1zM912.1 114.2c0 2.8.2 3 .9 1.3 1.2-2.8 1.2-4.5 0-4.5-.5 0-1 1.5-.9 3.2zM911.2 120.5c0 1.6.2 2.2.5 1.2.2-.9.2-2.3 0-3-.3-.6-.5.1-.5 1.8zM913 123c0 1.9.3 2.1 1.2 1.2.9-.9.9-1.5 0-2.4-.9-.9-1.2-.7-1.2 1.2zM910 126.9c-.6 1.2-.7 2.1-.1 2.1 1 0 2.4-3 1.7-3.7-.2-.3-.9.5-1.6 1.6zM913 137.8c0 9.4-1.2 15.2-3.2 15.2-1 0-1-.7.1-2.8.4-.9 1.1-4.5 1.5-8 .5-5.7.4-6.3-1.2-6-1.3.2-1.6 1.1-1.4 3.5.2 1.7-.1 3.4-.7 3.7-.6.4-.8 2-.5 3.5.4 1.6.1 3.7-.5 4.8-.9 1.5-1.1 1.3-1.2-1.7l-.1-3.5-1.3 3c-.8 1.6-1.2 3.7-.9 4.6.6 2.1-2.2 5-4.1 4.3-1.1-.4-1.5.1-1.5 1.7 0 1.3-.3 3.4-.6 4.6-.4 1.4-.2 2.3.5 2.3.6 0 1.4-1.2 1.7-2.6.5-1.9.9-2.2 1.5-1.2.7 1 .9 1 .9-.3 0-.9.9-2.6 2-3.9 1.1-1.3 2-2.9 2-3.7 0-.7.5-1.3 1.1-1.3.7 0 .9.9.5 2.2-.4 1.3-.9 4-1.1 6.1-.6 5.6 1.8 4.3 2.6-1.4.3-2.4.7-4 .7-3.5.1.5.9.3 1.7-.4 1.2-1 1.5-.9 1.5.8 0 1.1-.7 2.3-1.6 2.6-1.3.5-1.4.9-.5 2 .9 1.1.6 2-1.2 4.2-1.3 1.6-2.8 4-3.2 5.4-.4 1.4-1.5 3.4-2.5 4.5-2.1 2.4-3.3 6-3.2 9.8 0 1.7-.3 2.5-1 2.1-.8-.5-.9.3-.5 2.4.4 1.8.3 3.2-.1 3.2-2.3 0-5.2 4.3-5.2 7.8 0 2.8.3 3.3 1 2.2.5-.8 1-2.5 1-3.8 0-1.3.6-2.2 1.5-2.2.8 0 1.5-.8 1.5-1.8 0-1.4.2-1.4.9-.3.5.8.7 2.1.4 2.9-.3.8-.2 1.7.3 2.1.5.3 1.5 1.7 2.1 3.1 1.1 2.4 1.2 2.5 1.2.2.1-1.2.7-2.2 1.6-2.2 1.6 0 2.1-1.6 3-10.3.4-3.8.3-5.7-.4-5.7-1 0-1.5 2-1.2 4.2 0 .4-.6 2.3-1.5 4.3-.8 2.1-1.3 4-1 4.3.3.3.6-.1.6-.8 0-.7.2-1.1.5-.8.9.9-.6 3.8-2 3.8-1.5 0-2.4-5-1.4-7.6.4-1.1.8-4.3.8-7.3.1-3 .4-5.1.8-4.8.9 1 4.3-2.3 4.3-4.2 0-2.8 2.1-9.1 3-9.1.5 0 .4 1.5-.2 3.2-.5 1.8-1.2 4.2-1.5 5.3l-.5 2 1.6-2 1.7-2-.5 2c-1.2 4.7-2.7 21-2.7 27.7 0 2.9-1.3 3.8-2.2 1.5-.3-.6-.6.9-.6 3.5-.1 2.6-.5 4.9-1.1 5.3-.5.3-.9 7-.9 15.3.1 13.3.2 14.4 1.4 11.7 1.4-2.9 1.4-2.7 1.3 3.7-.1 3.7-.5 6.5-.9 6.2-.5-.2-1 1-1.3 2.8-1.5 9.7-1.6 12.7-.5 15.2 1 2 1 2.9.1 3.8-.7.7-1.3 3-1.3 5.3l-.2 4-.7-4.5c-.4-2.5-.5-5-.2-5.5.3-.7.1-.7-.6 0-1.1 1-2 23-1 24.1.3.3.9-1.2 1.3-3.4.5-3.3 1-3.9 2.5-3.5 1.8.5 1.8.7.5 2.6-.8 1.2-1.5 2.8-1.5 3.5 0 1.2.3 1.2 1.7.1 1.5-1.3 1.6-1.2.7.6-.8 1.9-1.3 34.6-.5 39 .6 3.3.1 29.9-.5 29.3-.6-.6-1.5-24.7-1.5-41.6 0-5-.5-9.2-1-9.2s-1 .6-1.1 1.2c-.5 3.1-.2 24.8.3 24.8.3 0 .3 2.6.1 5.7-.2 3.2.1 10.6.7 16.5.6 6.7.7 11.6.1 13-.8 2.1-.9 1.9-1-1.2 0-1.9-.5-10.3-1-18.5-1.9-29.1-2-32.1-2.2-40.8-.1-4.9-.6-9.1-1-9.3-1.3-.8-1.4 5.8-.2 10.9.6 2.4.4 2.6-1.2 2.1-1.5-.5-1.6-.4-.8.5.7.8.9 2.5.4 4.7-.4 1.9-.4 3.7.1 4 1.2.7 1.2 7.4 0 7.4-.6 0-1.4 1.9-1.9 4.2-.6 3.2-.9.1-.8-11.7 0-8.8-.1-28.5-.3-43.8-.1-15.3 0-28.1.3-28.4.6-.7 1 7.6.6 14.4-.1 3.6.1 4.3 1.7 4.3 1.6.1 1.6.2.3 1.3-1.5 1.2-2.5 7.1-1 6.2.4-.3 1 1.1 1.2 2.9.2 1.9 0 3.2-.4 3-.4-.3-.8.2-.8 1 0 .9.5 1.6 1 1.6.6 0 1 1.3 1 2.9 0 1.7-.4 3.3-1 3.6-1.5.9-1.2 5.1.4 6.4 1 .9 1.6.9 2.1.1 1-1.6-.6-43.1-1.6-42.4-1.5.9-1-1.6.6-3.2 1.2-1.2 1.4-3.4 1-10-.2-4.7-.7-9-1-9.5-.9-1.3.5-12 1.5-11.4.9.6.7-9.1-.2-10.1-.4-.4-.4-1.2.1-1.8 1.6-2.1 1.9-6.6.4-6.6-1.2 0-1.2-.3 0-2.1 1.5-2.1 1.2-5.9-.4-5.9-.5 0-.9 1.3-.9 2.9 0 1.9-.8 3.5-2 4.4-1.9 1.3-1.9 1.6-.5 5.7 1.8 5.2 1.8 5.1.1 4.4-.9-.3-1.5.5-2 2.7-1.1 5.6-.6 7.7.9 3.4 1.3-3.9 1.3-3.8.8 3.5-.4 4.6-1.2 8.2-2.1 9.3-.8.9-1.2 2-.8 2.4.4.4 1.1.1 1.5-.6 1.3-2 2.3 4.9 1.5 9.8-.5 3.1-1 3.9-1.9 3.1-2.3-1.9-3.5 15.2-3.3 49v20.5c0 1.6-.1 5-.1 7.5-.1 2.5.1 8.5.4 13.3.2 4.8.1 8.6-.3 8.3-.5-.3-.6.8-.4 2.5.3 2 0 2.9-.9 2.9-1 .1-1 .4.2 1.3 1.5 1.1 1.7 3.5 1.4 19-.1 9.8-.5 18-.9 18.4-.3.4-1.3-.2-2-1.3-3-4-2.7-.3.3 4 2.4 3.5 2.9 5 2.2 6.7-.7 2.1-.9 1.9-1.5-1.6-.6-3.2-.9-3.6-1.7-2.2-.6.9-.7 2.2-.3 2.8.4.8 0 .7-1.5-.1-1.2-.6-1.9-1.5-1.6-2 .3-.5.1-1.1-.4-1.4-.5-.4-1.2-.2-1.6.3-.3.6-1.2.8-2 .4-1-.3-1.5.1-1.5 1.6 0 1.4.5 1.9 1.5 1.5.8-.4 1.5-.1 1.5.5s.7.8 1.6.4c1.4-.5 1.5-.3.5 1.5-1.6 2.9-2.7 3.4-4.1 1.7-.7-.8-1.9-1.5-2.8-1.5-1.3 0-1.2.3.3 2 1.5 1.6 1.5 2.2.5 3.5-1 1.2-1 1.5.4 1.5.9 0 1.6-.5 1.6-1.1 0-.5.5-.7 1-.4 1.5 1 1.2 2.5-.7 3.2-1 .3-3.1 1.5-4.6 2.6l-2.8 1.9 3 1.9c3.1 1.9 4.1 3.9 1.9 3.9-.6 0-.8.3-.5.6.4.4 1.3.1 2.2-.6 1.1-.9 1.2-1.6.4-2.6-.8-1-.5-1.9 1-3.5l2.1-2.3v2.6c0 1.4.9 3.7 2 5.1 1.1 1.4 2 2.8 2 3.1 0 .3-2.5.4-5.7.3-6.1-.2-7.3.7-4.8 3.8 1.6 1.9 1.6 1.9 1-.3-.6-2.1-.5-2.2 1.3-1.1 2.9 1.8 5.6 7 5.4 10.2-.3 3 2.9 10 3.5 7.9.8-2.1 2.3-1.3 2.3 1.1 0 1.3.7 3.2 1.5 4.3 1.3 1.7 1.4 1.7 1.5.1 0-.9.4-1.7.9-1.7 1.4 0 .6-6.7-1-8.7-.8-1 .1-.5 1.9 1 1.9 1.5 3.8 2.4 4.4 2.1.6-.4.6-1-.1-1.7-.6-.6-1.1-2.2-1.1-3.6-.1-2.5-.1-2.6-1.6-.6-1.3 1.7-1.6 1.8-2.4.5-1.7-2.6-1.2-4.2 1-3.5 1.1.3 2 .2 2-.3s-.9-1.7-2-2.7c-1.1-1-2-2.3-1.9-2.9 0-.6.3-.6.6.1.2.7 1 1.3 1.7 1.3s2.3 1.8 3.6 3.9c1.4 2.1 2.9 3.7 3.4 3.4.4-.3 1.2.1 1.7.8.7 1.1.9 1 .9-.5s.3-1.6 1.5-.6c2 1.7 2-.5 0-4-.8-1.4-1.4-3.2-1.4-4 0-.9.5-.6 1.1.8 1.5 3.1 3 2.8 1.8-.4-.6-1.4-1-3.3-.9-4.2.2-4.2 2.8 5.9 2.8 11 .1 4.4.5 5.8 1.6 5.8 1.2 0 1.3-1.4.8-8.3-.3-4.5-1.4-9.9-2.3-11.9-1-2-1.5-4-1.2-4.3.3-.3 2 2.9 3.8 7.2 1.8 4.3 3.6 7.6 4.1 7.3 1.5-.9.2-9.1-1.9-11.8-1.2-1.5-1.1-1.5.4-.4.9.7 2.3 3.3 3.2 5.8 1.7 4.9 5.6 10.7 5.6 8.2 0-.9-.7-3.3-1.6-5.4-.9-2.1-1.6-5.3-1.6-7 0-2.5-.4-3.1-1.5-2.7-1.3.5-1.3.3-.1-1.6.9-1.6 1-2.4.2-2.9-.7-.4-1.4-2.5-1.6-4.7-.3-2.2-1-5-1.6-6.2-.7-1.1-1.1-3.8-1-6 .2-2.1-.1-6.1-.5-8.8-.4-3.3-.3-4.3.2-3 .9 1.8 1 1.7 1-.8.1-1.5-.3-2.7-.8-2.7s-1.2-.8-1.5-1.8c-.9-3.1-2.7-38.6-2.8-57.2-.1-14.7-.4-18-1.5-17.6-.7.3-1.3.1-1.3-.3 0-.5.5-1.1 1-1.3.6-.2 1.2-10.6 1.5-27.8.6-28.5.9-36.5 2.1-58.5.4-7.2.8-15.7 1-19 .4-10.5 1.4-23.9 1.9-27.1.8-4.6.9-44.4.1-45.1-.3-.3-.6 3.9-.6 9.5zm-6.6 40.7c-.4.8-.8 1.5-1 1.5-.2 0-.4-.7-.4-1.5s.4-1.5 1-1.5c.5 0 .7.7.4 1.5zm2.1 38.7c-1.3 13.4-1.4 14.1-2.6 10.4-.4-1.7-.5-3.5 0-4 .5-.6.8-2.5.8-4.1 0-3.5 1.1-8.5 1.9-8.5.3 0 .3 2.8-.1 6.2zm-1.8 19.5c-.3 1-.5.4-.5-1.2 0-1.7.2-2.4.5-1.8.2.7.2 2.1 0 3zm-.2 24.8c-.3 2.6-.9 4.5-1.2 4.2-.2-.3-.1-2.6.2-5.2.4-2.5 1-4.4 1.2-4.1.3.3.2 2.6-.2 5.1zm-10.5 1.9c0 .3-.4.8-1 1.1-.5.3-1 .1-1-.4 0-.6.5-1.1 1-1.1.6 0 1 .2 1 .4zm10.2 16.8c-.7.7-1.2.8-1.2.2 0-1.4 1.2-2.6 1.9-1.9.2.3-.1 1.1-.7 1.7zm-10.4 64.7c-.3 2-.8 4.4-1.2 5.4-.6 1.3-.5 1.6.8 1.1 1.8-.7 2.3 2 .6 3.1-.5.3-1 1.8-1 3.3 0 1.5-.5 3.4-1 4.2-1.1 1.7-.7-10.9.6-19.5.9-5.5 1.8-3.5 1.2 2.4zm.9 13.3c-.3.8-.6.5-.6-.6-.1-1.1.2-1.7.5-1.3.3.3.4 1.2.1 1.9zm-.1 9c-.7 5.8-2.6 10.1-2.6 6.1 0-3.2 2.1-12.5 2.7-11.9.3.3.2 2.9-.1 5.8zm8.1 3.5c-.3 1-.5.4-.5-1.2 0-1.7.2-2.4.5-1.8.2.7.2 2.1 0 3zm-10 9.5c-.3.7-.5.2-.5-1.2s.2-1.9.5-1.3c.2.7.2 1.9 0 2.5zM906 379c0 .5-.4 1-1 1-.5 0-1-.5-1-1 0-.6.5-1 1-1 .6 0 1 .4 1 1zm-1.4 7.7c-.8 1.5-1.5 2.2-1.6 1.5 0-1.4 2.2-5.3 2.8-4.7.2.2-.3 1.6-1.2 3.2zM906 400c0 .5-.4 1-1 1-.5 0-1-.5-1-1 0-.6.5-1 1-1 .6 0 1 .4 1 1zm-16 9c0 .5-.2 1-.4 1-.3 0-.8-.5-1.1-1-.3-.6-.1-1 .4-1 .6 0 1.1.4 1.1 1zm13.5.1c1.1 1.7-.1 2.1-1.5.4-.8-.9-.8-1.5-.2-1.5.6 0 1.4.5 1.7 1.1zm2.5 5.5c0 .8-.4 1.2-1 .9-.5-.3-1-1-1-1.6 0-.5.5-.9 1-.9.6 0 1 .7 1 1.6zm-22.1 3.1c0 2.5-.1 2.5-.8.8-.5-1.1-.7-2.3-.4-2.8.8-1.4 1.3-.7 1.2 2zm.1 12.8c1.6 1.9.1 1.9-2.4 0-1.6-1.2-1.7-1.5-.4-1.5.8 0 2.1.7 2.8 1.5zm17.9 4.5c1.1 2.4 1.1 2.4-.3.6-.9-1.1-1.6-2.4-1.6-3 0-1.2.5-.7 1.9 2.4zM577 131c-1.3.8-1.2 1 .3 1 .9 0 1.7-.5 1.7-1 0-1.2-.1-1.2-2 0zM588 131.1c0 .5.5.7 1 .4.6-.3 1-.8 1-1.1 0-.2-.4-.4-1-.4-.5 0-1 .5-1 1.1z" />
-            <path d="M908.5 133c-1.6 3.1-1.6 3.1.4 2.1 2.7-1.5 3.5-3.1 2.3-4.3-.7-.7-1.5 0-2.7 2.2zM570.6 132.5c-1.5 1.1-1.6 1.5-.5 1.5.8 0 2.4-.7 3.5-1.5 1.4-1.1 1.5-1.4.4-1.4-.8 0-2.4.6-3.4 1.4zM550 138c0 .5.7.7 1.5.4.8-.4 1.5-.8 1.5-1 0-.2-.7-.4-1.5-.4s-1.5.4-1.5 1zM541.5 141c-.3.5.1 1 .9 1s1.8-.5 2.1-1c.3-.6-.1-1-.9-1s-1.8.4-2.1 1zM634 144.5c-.7.9-.8 1.5-.2 1.5.5 0 1.2-.7 1.6-1.5.3-.8.4-1.5.2-1.5s-.9.7-1.6 1.5zM881.1 148.9c-.1 2.7.2 3.2 1.1 2.3.8-.8.8-1.7 0-3.4-1-2.2-1.1-2.1-1.1 1.1zM528 149c0 .5.7.7 1.5.4.8-.4 1.5-.8 1.5-1 0-.2-.7-.4-1.5-.4s-1.5.4-1.5 1zM523.5 151c-.3.5-.2 1 .4 1 .5 0 1.3-.5 1.6-1 .3-.6.2-1-.4-1-.5 0-1.3.4-1.6 1zM885.4 152.1c-.3.6-.1 1.6.5 2.2.8.8 1.1.5 1.1-1.1 0-2.4-.5-2.8-1.6-1.1zM903 165c0 .5.5 1 1 1 .6 0 1-.5 1-1 0-.6-.4-1-1-1-.5 0-1 .4-1 1zM900.4 168c-1.5.6-2.6 2.8-4.3 9.6-1.2 4.9-2.3 10.4-2.4 12.1 0 2-.5 3.3-1.3 3.3-2 0-2.5 3.6-4 28.4-.5 7.5-.4 10.7.4 10.3 1.1-.8 2.4-9.9 2.7-19.8.1-3.6.4-7 .8-7.5.3-.6.2-1.6-.3-2.4-.9-1.4.6-5.8 2.6-7.5 1.9-1.6 5.7-12.7 5.8-17 .1-2.2.7-5.3 1.4-6.9 1.5-3.4 1.4-3.6-1.4-2.6zM890.1 181.7c-1.4 4.9-1.4 6.2.2 7.7.9 1 1.2 0 1.4-5.1.2-3.5.1-6.3-.2-6.3-.2 0-.9 1.7-1.4 3.7zM885.1 182.2c0 2.8.2 3 .9 1.3 1.2-2.8 1.2-4.5 0-4.5-.5 0-1 1.5-.9 3.2zM1029.2 181c0 1.4.2 1.9.5 1.2.2-.6.2-1.8 0-2.5-.3-.6-.5-.1-.5 1.3zM1096 179.9c0 .5.7 1.4 1.5 2.1 1.6 1.4 1.8 1 .9-1.4-.6-1.7-2.4-2.2-2.4-.7zM907 183c0 .5.5 1 1 1 .6 0 1-.5 1-1 0-.6-.4-1-1-1-.5 0-1 .4-1 1zM1142.5 192.1c-1 1.5.3 2.5 1.6 1.2.7-.7.7-1.3.1-1.7-.6-.3-1.4-.1-1.7.5zM884.7 195.7c-.4.3-.7 1.1-.7 1.7 0 .6.5.5 1.2-.2.6-.6.9-1.4.7-1.7-.3-.3-.9-.2-1.2.2zM1126.4 198.5c-1.6 1.2-1.7 1.4-.3 1.5.9 0 1.9-.7 2.3-1.5.7-1.9.5-1.9-2 0zM824 203.5c0 1.9.5 3.5 1 3.5.6 0 1-1.6 1-3.5s-.4-3.5-1-3.5c-.5 0-1 1.6-1 3.5zM883 203.4c-1.2 3.3-1.2 3.6-.1 3.6.9 0 2.2-5.2 1.5-5.9-.2-.2-.8.8-1.4 2.3zM1110.9 207.7c-1.9 2.1-2.1 2.5-.6 2 1.1-.4 3-1 4.3-1.3 2-.5 2.2-.9 1.3-2-1.6-1.9-2.2-1.8-5 1.3zM882.7 211c-.3 1.1-.1 2 .4 2s.9-.9.9-2-.2-2-.4-2-.6.9-.9 2zM900.6 211.6c-.4.9-.4 1.9-.2 2.1.8.8 2.6-.8 2.6-2.3 0-1.9-1.6-1.8-2.4.2zM566 216c0 .5.5 1 1 1 .6 0 1-.5 1-1 0-.6-.4-1-1-1-.5 0-1 .4-1 1zM900.7 221.7c-.4.3-.7 1.6-.7 2.7.1 2.1.1 2.1 1.5.2.8-1.1 1.5-2.3 1.5-2.8 0-1-1.4-1.1-2.3-.1zM882 227.6c0 .8.5 1.2 1 .9.6-.3 1-1 1-1.6 0-.5-.4-.9-1-.9-.5 0-1 .7-1 1.6zM892.1 231.6c0 1.1.3 1.4.6.6.3-.7.2-1.6-.1-1.9-.3-.4-.6.2-.5 1.3zM888.8 233.6c-.3.5-.5 6.7-.6 13.9-.1 9.4.2 12.3.9 10.6.5-1.3.9-5 .8-8.2-.1-3.5.3-6.5 1.1-7.4 1-1.2 1-1.8 0-3-.6-.8-.9-1.9-.5-2.4.3-.6.2-1.8-.3-2.7-.5-.9-1.1-1.3-1.4-.8zM900 234c-.2.8-.4 1.9-.5 2.5-.1.5-.6 1.8-1 2.9-.5 1.1-.3 2.6.4 3.8.7 1.1 1 3.2.6 4.9-.4 1.9-.3 2.9.5 2.9s.9.7.5 1.7c-1.3 3.2-1.8 17.3-.7 17.3 1.4 0 1.9-2.2 2.1-10 .2-3.5.5-6.8.8-7.3.3-.5.1-1.2-.4-1.6-.6-.3-.8-2.7-.3-5.8.3-2.9.3-5.6-.1-6-.4-.5-.9-2.2-1.2-3.8-.2-1.7-.5-2.3-.7-1.5zM228.3 238.2c-1.8 1.3-3.3 2.7-3.3 3.1 0 .4 1.6-.4 3.5-1.9 3.1-2.3 4.1-3.4 3.2-3.4-.1 0-1.7 1-3.4 2.2zM1270.9 240.6c-1.7 4.6-1.9 6.6-.6 6.1.7-.2 1.2-.8 1.3-1.3 0-.5.4-2.2.9-3.7 1-3.8-.2-4.7-1.6-1.1zM882 245.5c0 .9-.7 1.8-1.5 2.1-.8.4-1.5 1.4-1.5 2.3 0 1.4.2 1.4 1 .1.8-1.2 1.2-1.3 1.9-.2.6 1.1 1 1 1.5-.3.9-2.5.7-5.5-.4-5.5-.5 0-1 .7-1 1.5zM903.6 246.9c-.3.5.1 1.4.8 2 1.2 1 1.3.8.9-.8-.6-2.3-.9-2.5-1.7-1.2zM881.8 257c-1.2 3.8-1 6.6.3 5.3.6-.6.9-2.6.7-4.4-.3-3.2-.3-3.3-1-.9zM429.7 262.4c-.4 1-.1 1.3.9.9 1.5-.6 1.9-2.3.5-2.3-.5 0-1.1.7-1.4 1.4zM881.7 264.7c-1 .9-.9 10.1.1 10.7.4.2.8-2.2.8-5.5 0-3.2 0-5.9-.1-5.9s-.5.3-.8.7zM886 278.5c0 1.5.4 2.4 1 2 .6-.3 1-1.2 1-2s-.4-1.7-1-2c-.6-.4-1 .5-1 2zM880.9 278.2c-.6 6.6-.6 7.8.2 7.8.5 0 .9-2 .9-4.5s-.2-4.5-.5-4.5-.5.6-.6 1.2zM1263.5 279.7c-.5 1-.7 2.2-.6 2.8.1.5.2 6.5.2 13.3-.1 7.3.3 12.1.8 11.7.6-.3.7 2.6.4 7.2-.7 9.6.5 9.2 1.6-.5.5-5.5.4-7.3-.6-7.6-.7-.3-1.3-1.6-1.3-2.8 0-1.6.3-1.9.9-1 .7 1.1 1 1 1.5-.3.8-2.2.8-2.7-.2-2.1-.5.3-.9-4.5-1-10.6 0-6.1-.3-11.2-.5-11.5-.3-.2-.8.4-1.2 1.4zM1268 282c0 .5.5 1 1.1 1 .5 0 .7-.5.4-1-.3-.6-.8-1-1.1-1-.2 0-.4.4-.4 1zM1268 285.4c0 .9.5 1.6 1 1.6.6 0 1-.4 1-.9 0-.6-.4-1.3-1-1.6-.5-.3-1 .1-1 .9zM380.3 287.8c-4.1 1.5-5.7 3.2-3 3.2 2.5 0 8.7-2.1 8.7-3 0-1.7-1.5-1.7-5.7-.2zM356.7 288.6c-1.5.8-4.7 1.6-7 1.6-4.6.2-8.7 2.2-8.7 4.4 0 1.6 1.6 1.9 2.5.4.3-.6 2.6-1 5-1 5.3 0 9-1.5 11-4.7 1.8-2.7 1.1-2.8-2.8-.7zM878 289.7c0 2.2.2 2.5 1 1.3 1.3-2 1.3-4 0-4-.5 0-1 1.2-1 2.7zM1258.7 290c-.3 1.1-.1 2 .4 2s.9-.9.9-2-.2-2-.4-2-.6.9-.9 2zM1268.7 292.9c-.2 1-.1 2.5.4 3.2.6 1 .9.4.9-1.9 0-3.6-.5-4.2-1.3-1.3zM358.5 293c-.3.5-.1 1 .4 1 .6 0 1.1-.5 1.1-1 0-.6-.2-1-.4-1-.3 0-.8.4-1.1 1zM881.6 294.6c-.3.9-.6 2.3-.6 3.2.1 1.5.3 1.4 1.5-.2 1.7-2.2 2-4.6.6-4.6-.5 0-1.2.7-1.5 1.6zM369.5 296c-.3.6.5 1 1.9 1s2.8-.5 3.1-1c.3-.6-.5-1-1.9-1s-2.8.4-3.1 1zM1259 297.2c0 1.2.3 1.8.7 1.4 1.1-1 1.2-3.6.2-3.6-.5 0-.9 1-.9 2.2zM348.1 297.1c-1.6.6-2.1 1.3-1.4 2 1.3 1.3 4.3.1 4.3-1.7 0-.8-.1-1.4-.2-1.3-.2 0-1.4.5-2.7 1zM364 297c0 .5.6 1 1.4 1 .8 0 1.8-.5 2.1-1 .3-.6-.3-1-1.4-1-1.2 0-2.1.4-2.1 1zM334 299c-1.2.8-.9 1 1.3 1 1.5 0 2.7-.5 2.7-1 0-1.3-2-1.3-4 0zM341 299c0 .5.9 1 2 1s2-.5 2-1c0-.6-.9-1-2-1s-2 .4-2 1zM402 299.1c0 .5.5.7 1 .4.6-.3 1-.8 1-1.1 0-.2-.4-.4-1-.4-.5 0-1 .5-1 1.1zM901.2 299.3c-1.3 1.5-1.4 13.7-.1 13.7.5 0 .9-.3 1-.8.1-1.5.4-14.2.3-14.2-.1 0-.6.6-1.2 1.3zM1259.6 301.2c-1.2 3.8-1.3 5.7-.2 6.1.7.3 1.1-1 1.1-3.4.1-4-.3-4.9-.9-2.7zM164 302c0 .5.5 1 1 1 .6 0 1-.5 1-1 0-.6-.4-1-1-1-.5 0-1 .4-1 1zM196.3 302.3c-1.8.6-3.3 1.5-3.3 1.9 0 1.2 1.4 1 6.1-1.2 4.6-2 2.7-2.5-2.8-.7zM368.7 301.7c-1.2 1.2-.7 2.2.8 1.7.8-.4 1.5-1 1.5-1.5 0-1.1-1.4-1.2-2.3-.2zM384.8 301.7c.7.3 1.6.2 1.9-.1.4-.3-.2-.6-1.3-.5-1.1 0-1.4.3-.6.6zM1268.7 303c.3 1.1.7 2 .8 2 .1 0 .5-.9.8-2 .3-1.2 0-2-.8-2s-1.1.8-.8 2zM351.8 302.6c-.3.3-3.4.6-6.9.7-3.4.1-6.4.7-6.6 1.2-.2.8 9.2.6 18-.4.4-.1.7-.6.7-1.1 0-1-4.3-1.4-5.2-.4zM358.5 303c-.4.6.6 1 2.4 1 1.6 0 3.3-.5 3.6-1 .4-.6-.6-1-2.4-1-1.6 0-3.3.4-3.6 1zM882.2 303.2c-.7.7-1.2 2.8-1.2 4.7 0 3.5 0 3.5 1.6 1.3.9-1.3 1.3-2.9 1.1-3.6-.3-.8 0-1.9.6-2.5.6-.6.7-1.1.1-1.1-.5 0-1.5.5-2.2 1.2zM321.8 303.7c-3 .8-2 2.6 1.2 2.1 3-.4 4.2.6 1.8 1.5-.7.2-1 .7-.7 1.1.4.4 1.3.1 2.1-.7.8-.8 2.4-1.8 3.6-2.1 1.2-.4 2.2-1.2 2.2-1.8 0-.5-.7-.7-1.5-.4-.8.3-2.2.9-3 1.2-.8.3-1.5.1-1.5-.5 0-1.1-1.1-1.2-4.2-.4zM158 305c0 .5.7 1 1.5 1s1.5-.5 1.5-1c0-.6-.7-1-1.5-1s-1.5.4-1.5 1zM306.5 305.1c-.4.6-.5 1.3-.2 1.6.7.7 2.7-.6 2.7-1.8 0-1.2-1.7-1.2-2.5.2zM380 305c0 .5.7 1 1.5 1s1.5-.5 1.5-1c0-.6-.7-1-1.5-1s-1.5.4-1.5 1zM299 306c-1.3.8-1.3 1 .2 1 .9 0 2-.5 2.3-1 .8-1.3-.5-1.3-2.5 0zM1268.8 306.7c-.2.5-.1 2.6.3 4.8l.6 4 .2-4.8c.1-4.4-.3-5.7-1.1-4zM309 309c-2 1.2-.9 2.1 1.8 1.3 1.2-.3 1.9-.9 1.6-1.4-.7-1.1-1.5-1.1-3.4.1zM343.3 308.7c.9.2 2.3.2 3 0 .6-.3-.1-.5-1.8-.5-1.6 0-2.2.2-1.2.5zM463 309c-1.3.8-1.3 1 .2 1 .9 0 2-.5 2.3-1 .8-1.3-.5-1.3-2.5 0zM321.8 310.7c1.2.2 3 .2 4 0 .9-.3-.1-.5-2.3-.4-2.2 0-3 .2-1.7.4zM396.8 311.7c1.2.2 3 .2 4 0 .9-.3-.1-.5-2.3-.4-2.2 0-3 .2-1.7.4zM381.9 313.1c-1 .6-.7.9 1.4.9 2.9 0 3.5-.5 1.4-1.3-.8-.3-2-.1-2.8.4zM374.5 314c-1.6.7-1.7.9-.3.9.9.1 2.3.4 3.1.7.8.3 1.7-.2 2-1 .7-1.8-1.5-2-4.8-.6zM340.5 317c-.3.5-.1 1 .4 1 .6 0 1.1-.5 1.1-1 0-.6-.2-1-.4-1-.3 0-.8.4-1.1 1zM448.2 317.6c-1 1.2-1 1.4.2 1.4 1.7 0 3-1.6 1.9-2.3-.5-.3-1.4.1-2.1.9zM374.3 318.8c-3 .2-4.6.9-5.3 2.3-1.3 2.2 0 2.5 3.3.7 1.2-.6 3.5-1.4 5.2-1.8 4.1-.8 2-1.6-3.2-1.2zM363.5 320.1c-.8 1.2 1.4 3.1 3 2.6.6-.2 1-1 1-1.8 0-1.7-3.1-2.4-4-.8zM305.8 320.7c.7.3 1.6.2 1.9-.1.4-.3-.2-.6-1.3-.5-1.1 0-1.4.3-.6.6zM357.4 321.5c-1.6 1.2-1.7 1.4-.2 1.5 2.3 0 4.8-1.1 4.8-2.1 0-1.4-2.4-1.1-4.6.6zM333 322c-.9.6-1 1-.3 1 .6 0 1.5-.5 1.8-1 .8-1.2.4-1.2-1.5 0zM1259.8 323.4c-.3 1.4 0 2.7.6 3 .7.3.6.8-.4 1.5-1.8 1.3-.1 1.5 1.8.2 1.6-1 1.2-6.3-.4-6.8-.6-.2-1.4.7-1.6 2.1zM1274.2 324c0 1.9.2 2.7.5 1.7.2-.9.2-2.5 0-3.5-.3-.9-.5-.1-.5 1.8zM324.5 322.9c-1 1.7.9 1.9 4 .6l3-1.3-3.2-.1c-1.7-.1-3.4.3-3.8.8zM341.5 323c.3.5 1.3 1 2.1 1s1.4-.5 1.4-1c0-.6-.9-1-2.1-1-1.1 0-1.7.4-1.4 1zM359 325c0 .5.5 1 1 1 .6 0 1-.5 1-1 0-.6-.4-1-1-1-.5 0-1 .4-1 1zM882.1 328.2c.1 1.8.4 4.4.9 5.8l.7 2.5.8-2.8c.4-1.7.3-3.3-.4-4.1-.8-1-.9-1.6-.1-2.1 1.5-.9 1.2-2.5-.5-2.5-1 0-1.5 1-1.4 3.2zM1264.1 326.6c0 1.1.3 1.4.6.6.3-.7.2-1.6-.1-1.9-.3-.4-.6.2-.5 1.3zM274 327l-3.5.7 4.3.2c2.7 0 4.1-.3 3.7-.9-.3-.6-.7-.9-.8-.9-.1.1-1.8.5-3.7.9zM361 328c-1.1.7-.4 1 2.7 1 2.3 0 4.5-.5 4.8-1 .8-1.3-5.5-1.3-7.5 0zM240 329c0 .5.4 1 .9 1 .6 0 1.3-.5 1.6-1 .3-.6-.1-1-.9-1-.9 0-1.6.4-1.6 1zM226.5 330c-.4.6 1.1 1 3.9 1 2.5 0 4.8-.5 5.1-1 .4-.6-1.1-1-3.9-1-2.5 0-4.8.4-5.1 1zM1273.8 333c-.2 2.2-.1 4 .3 4s.9-1.8 1.1-4c.2-2.2.1-4-.3-4s-.9 1.8-1.1 4zM1264.1 335.7c.1 5.3.6 4.9 1.1-1 .2-2 0-3.7-.4-3.7s-.8 2.1-.7 4.7zM1260.9 333.2c-.7 1.4-.6 5.9.2 8.5.8 2.6.7 3.7-.5 5.5-1.9 2.6-2.1 4.2-.5 3.2.8-.4.8 0 0 1.5-.9 1.6-.9 2.1.1 2.1.9 0 .9.3-.2 1-1.2.8-1.2 1-.1 1 .9 0 1.1.7.7 2.2-1.4 4.6-1.5 5.8-.5 5.8.5 0 .9-1.1.9-2.4 0-1.4.5-2.8 1-3.1 1.3-.8 1.3-10.5-.1-10.5-.6 0-.7-.5-.2-1.3 1.1-1.7 1.3-13.7.2-13.7-.5 0-1 .1-1 .2zM875.6 337.9c-.3 2.2-.2 5 .3 6.3.8 1.8 1 1 1-4 .1-3.4-.1-6.2-.3-6.2-.2 0-.6 1.8-1 3.9zM1264.1 342.6c0 1.1.3 1.4.6.6.3-.7.2-1.6-.1-1.9-.3-.4-.6.2-.5 1.3zM1269 342c0 .5.5 1 1.1 1 .5 0 .7-.5.4-1-.3-.6-.8-1-1.1-1-.2 0-.4.4-.4 1zM883.5 344c-.3.6.1.7.9.4 1.8-.7 2.1-1.4.7-1.4-.6 0-1.3.4-1.6 1zM1268.6 351.2c-.9 3.6-.7 8.8.4 8.8.6 0 1-2.5 1-5.5 0-5.9-.4-6.9-1.4-3.3zM1271.1 352.6c0 1.1.3 1.4.6.6.3-.7.2-1.6-.1-1.9-.3-.4-.6.2-.5 1.3zM1263.7 357.5c-.4.8-.2 1.7.3 2 .6.4 1-.3 1-1.4 0-2.5-.6-2.7-1.3-.6zM1263.6 362.5c-.3.8-.1 1.5.4 1.5.6 0 1-.7 1-1.5s-.2-1.5-.4-1.5-.6.7-1 1.5zM1268 365.8c0 2.6.4 4.1 1 3.7 1.3-.8 1.3-4.5 0-6.5-.7-1.1-1-.3-1 2.8zM886.5 365.2c-1 2.6-.2 4.3 1.2 2.6.7-.8.8-1.9.3-2.8-.8-1.2-1-1.2-1.5.2zM1259.9 365.9c1.1.7 1.2 2.2.7 6.2-.4 3-.9 6.6-1.1 8.1-.2 1.6 0 2.8.6 2.8.5 0 .3 1.5-.5 3.4-.9 2.3-1.1 3.7-.4 4.4.7.7.6 1.7-.1 3.1-1.4 2.5-.3 2.8 1.5.3.8-1 1.3-4.3 1.4-7.9 0-3.9.4-6.3 1.3-6.6.7-.3.5-.6-.6-.6-1.3-.1-1.6-.7-1.3-2.4.3-1.2.8-4.3 1.1-7 .6-4.4.5-4.7-1.7-4.7-1.7 0-1.9.3-.9.9zM1264.3 370c0 3 .2 4.3.4 2.7.2-1.5.2-3.9 0-5.5-.2-1.5-.4-.2-.4 2.8zM1273.2 371.5c0 1.6.2 2.2.5 1.2.2-.9.2-2.3 0-3-.3-.6-.5.1-.5 1.8zM885.7 372.6c-.7 3.6-.1 4.2 1.8 2.3 1.3-1.3 1.3-1.8.2-3.4-1.3-1.7-1.4-1.6-2 1.1zM1268.1 377.7c.2 7 .6 6.3 1.1-2 .2-2.6 0-4.7-.4-4.7-.5 0-.7 3-.7 6.7zM1270.4 373.4c-.3.8-.4 2.7-.2 4.3.3 2.3.5 2 .9-1.5.5-4.3.2-5.3-.7-2.8zM888 376.5c-.1.5-.2 1.4-.1 1.8.1.4-.6 1-1.4 1.3-1.4.6-2.2 3.4-.9 3.4 1.5 0 3.5-3.3 3.1-5.2-.2-1.3-.5-1.9-.7-1.3zM885.5 386.6c0 1.3.3 2.4.8 2.4.4 0 .7-.9.7-1.9 0-1.1-.3-2.2-.7-2.4-.5-.3-.8.6-.8 1.9zM1263.2 389c0 1.4.2 1.9.5 1.2.2-.6.2-1.8 0-2.5-.3-.6-.5-.1-.5 1.3zM880 391c0 .5.5 1 1 1 .6 0 1-.5 1-1 0-.6-.4-1-1-1-.5 0-1 .4-1 1zM1272.1 391.6c0 1.1.3 1.4.6.6.3-.7.2-1.6-.1-1.9-.3-.4-.6.2-.5 1.3zM1259.6 398.2c-1.2 1.6-1.3 2.4-.4 2.9.8.5.9 1.1.1 2.1-1.4 1.8-2.3 9.2-1.2 10.9.6.9.9-.4.9-3.8 0-4.7.2-5.2 2.8-6.1 1.5-.6 2-1.1 1-1.1-1.1-.1-1.5-.7-1.1-2.1.3-1.1.9-2 1.4-2 .5 0 .9-.7.9-1.6 0-1.4-.2-1.5-1.7-.3-1.1 1-1.4 1-.9.1 1.3-2.1-.2-1.3-1.8 1z" />
-            <path d="M1261.1 406.9c-1 .6-.9 1 .3 1.4.8.3 1.8.2 2.1-.3.8-1.3-.9-2-2.4-1.1zM870.2 408.1c.2.7 1 1.5 1.7 1.7.8.3 1.2-.1.9-.9-.2-.7-1-1.5-1.7-1.7-.8-.3-1.2.1-.9.9zM869 410.8c0 .5.7 1.5 1.5 2.2.8.7 1.5.8 1.5.2 0-.5-.7-1.5-1.5-2.2-.8-.7-1.5-.8-1.5-.2zM1261.8 411.7c-.2 1-.1 3.4.2 5.3.6 3.3.7 3.2 1.1-1.8.4-4.9-.4-7.2-1.3-3.5zM1257.6 419.2c-.9 3.5-.8 3.8 1.5 3.8 2 0 2 .4 1.5 10.6-.4 6.3-.2 10.3.3 10 .5-.4 1.2 0 1.5.6.3.7.3.1 0-1.5-.4-1.6-.2-2.7.5-2.7.6 0 1.1-1 1-2.3 0-2.1-.1-2.1-.9-.2-.6 1.3-.8-1-.4-6.3.3-6.2.1-8.2-.8-8.2s-.8-.4.2-1.1c1.2-.8 1-.9-.9-.4-2.2.7-2.3.5-1.7-1.9.3-1.4.2-2.6-.3-2.6s-1.1 1-1.5 2.2z" />
-            <path d="M1257.7 424.7c-1.1 1.9-1.7 7.2-.9 7.7.6.3 1.2-1.3 1.4-3.7.5-4.3.4-5.4-.5-4zM868.2 426.1c.2.7 1 1.5 1.7 1.7.8.3 1.2-.1.9-.9-.2-.7-1-1.5-1.7-1.7-.8-.3-1.2.1-.9.9zM916 436.5c0 .8.4 1.5 1 1.5.5 0 .7-.7.4-1.5-.4-.8-.8-1.5-1-1.5-.2 0-.4.7-.4 1.5zM880.1 437.6c0 1.1.3 1.4.6.6.3-.7.2-1.6-.1-1.9-.3-.4-.6.2-.5 1.3zM903.9 444.2c-.1.7 0 1.7.1 2.3.1.5.4 2.1.6 3.5.9 5.2 1.5 7 2.3 7 1.2 0 2.1-3.9 1.1-4.5-.5-.3-1.1-2.3-1.5-4.5-.7-3.9-2.4-6.4-2.6-3.8zM910 447.5c0 2 3.2 8.5 4.2 8.5 1.4 0 .8-3.4-1.2-6.8-2.2-3.5-3-4-3-1.7zM896.7 447.6c-1.1 1.1-.8 6.2.4 6.9 1.7 1.1 2 .1 1.2-3.9-.5-2.1-1.2-3.4-1.6-3zM1259.9 451.2c-.1.2-.3 1.2-.4 2.3-.1 1.1-.5 3.1-1 4.5-.5 1.7-.4 2.2.3 1.5 1.2-1.1 2.7-8.5 1.8-8.5-.3 0-.7.1-.7.2zM1258 464.9c0 .7-.5.9-1 .6-1.6-1-2.5 2.6-.9 3.7 1.1.7 1.1 1.1.2 1.5-.7.2-1.3 1.4-1.3 2.4 0 1.1-.4 1.7-1 1.4-.5-.3-1 .1-1 1 0 1.2.4 1.3 1.8.6 1.8-1 4.9-10.7 3.8-11.8-.3-.4-.6-.1-.6.6z" />
-          </svg>
-        </Link>
-      </div>
-      <div className="side-wrapper">
-        <div className="side-title">Pages</div>
-        <div className="side-menu">
-          <Link style={navStyle} to="/about">
-            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20">
+      <Link to="/" style={navStyle}>
+        <NavIcon />
+      </Link>
+
+      <div className="mt-5">
+        <div className="text-zinc-400 mb-4">Pages</div>
+        <div className="">
+          <Link style={navStyle} to="/about" className={linkClassName}>
+            <svg class="fill-slate-50 w-5 h-5" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20">
               <path d="M.7 9.3l4.8-4.8 1.4 1.42L2.84 10l4.07 4.07-1.41 1.42L0 10l.7-.7zm18.6 1.4l.7-.7-5.49-5.49-1.4 1.42L17.16 10l-4.07 4.07 1.41 1.42 4.78-4.78z" />
             </svg>
             <h4>About</h4>
           </Link>
-          <Link style={navStyle} to="/search">
-            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20">
+          <Link style={navStyle} to="/search" className={linkClassName}>
+            <svg class="fill-slate-50 w-5 h-5" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20">
               <path d="M12.9 14.32a8 8 0 111.41-1.41l5.35 5.33-1.42 1.42-5.33-5.34zM8 14A6 6 0 108 2a6 6 0 000 12z" />
             </svg>
             <h4>Search</h4>
@@ -40,11 +34,11 @@ function Nav({ isLoggedIn, seteIfUpdateMySentencePage }) {
         </div>
       </div>
 
-      <div className="side-wrapper">
-        <div className="side-title">Account</div>
+      <div className="mt-5">
+        <div className="text-zinc-400 mb-4">Account</div>
         <div className="side-menu">
           {isLoggedIn ? (
-            <Link style={navStyle} to="/mysentences">
+            <Link style={navStyle} to="/mysentences" className={linkClassName}>
               <h4>My Setences</h4>
             </Link>
           ) : (
@@ -54,8 +48,8 @@ function Nav({ isLoggedIn, seteIfUpdateMySentencePage }) {
           {isLoggedIn ? (
             ''
           ) : (
-            <Link style={navStyle} to="/login">
-              <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20">
+            <Link style={navStyle} to="/login" className={linkClassName}>
+              <svg class="fill-slate-50 w-5 h-5" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20">
                 <path d="M5 5a5 5 0 0110 0v2A5 5 0 015 7V5zM0 16.68A19.9 19.9 0 0110 14c3.64 0 7.06.97 10 2.68V20H0v-3.32z" />
               </svg>
               <h4>Log in</h4>
@@ -65,12 +59,12 @@ function Nav({ isLoggedIn, seteIfUpdateMySentencePage }) {
           {isLoggedIn ? (
             ''
           ) : (
-            <Link style={navStyle} to="/register">
+            <Link style={navStyle} to="/register" className={linkClassName}>
               {/* <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20"><path d="M10 12a6 6 0 110-12 6 6 0 010 12zm0-3a3 3 0 100-6 3 3 0 000 6zm4 2.75V20l-4-4-4 4v-8.25a6.97 6.97 0 008 0z"/></svg> */}
-              <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20">
-                <path d="M10 3.22l-.61-.6a5.5 5.5 0 00-7.78 7.77L10 18.78l8.39-8.4a5.5 5.5 0 00-7.78-7.77l-.61.61z" />
+              <svg class="fill-slate-50 w-5 h-5" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 122.88 122.88" >
+                <path d="M109.72 61.44v-.01h.02c0-13.31-5.42-25.38-14.17-34.13-8.75-8.74-20.81-14.16-34.1-14.16v.02h-.04v-.02c-13.3 0-25.38 5.42-34.13 14.17-8.74 8.75-14.16 20.81-14.16 34.11h.02v.04h-.02c0 8.88 2.41 17.22 6.59 24.36 8.38 14.31 23.92 23.92 41.68 23.93v-.02h.04v.02c13.31 0 25.38-5.42 34.13-14.17 1.44-1.44 2.78-2.95 4-4.52a48.149 48.149 0 0 0 10.14-29.62zM39.91 86.16V35.18h26.26c4.87 0 8.59.42 11.16 1.26 2.58.83 4.65 2.37 6.23 4.64 1.58 2.26 2.36 5.01 2.36 8.25 0 2.84-.61 5.27-1.8 7.32-1.21 2.05-2.87 3.71-4.98 5-1.35.81-3.19 1.48-5.53 2 1.88.63 3.24 1.26 4.11 1.89.57.42 1.41 1.31 2.52 2.67 1.1 1.38 1.84 2.42 2.21 3.17l7.62 14.78H72.26l-8.42-15.59c-1.07-2-2.02-3.32-2.86-3.93-1.13-.78-2.42-1.18-3.86-1.18h-1.39v20.69H39.91v.01zm15.83-30.33h6.65c.71 0 2.1-.23 4.17-.69 1.04-.21 1.9-.75 2.56-1.6.65-.85.99-1.84.99-2.95 0-1.65-.52-2.91-1.57-3.79-1.05-.89-3-1.33-5.88-1.33h-6.92v10.36zm67.11 5.6h.03v.01c0 14.19-4.82 27.27-12.92 37.68a62.403 62.403 0 0 1-5.08 5.76c-11.1 11.1-26.45 17.97-43.43 17.98v.03h-.04v-.03c-22.59-.01-42.34-12.22-52.99-30.4-5.33-9.1-8.39-19.7-8.39-31H0v-.04h.03C.03 44.44 6.9 29.09 18 18 29.09 6.9 44.45.03 61.43.03V0h.04v.03C78.44.04 93.79 6.9 104.88 18c11.1 11.09 17.97 26.45 17.97 43.43z"/>
               </svg>
-              <h4>Register</h4>
+              <h4>Create</h4>
             </Link>
           )}
         </div>
