+++ src/index.css
@@ -1,162 +1,19 @@
-@import url('https://fonts.googleapis.com/css2?family=Rubik:wght@300&display=swap');
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
 
-@media (min-width: 1200px){
-  body {
-    margin: 0;
-    /* background-image: url(https://4kwallpapers.com/images/wallpapers/macos-big-sur-apple-layers-fluidic-colorful-dark-wwdc-2020-5120x2880-1432.jpg); */
-    /* background-image: url(https://4kwallpapers.com/images/wallpapers/bannalpsee-switzerland-aurora-borealis-starry-sky-landscape-1920x1080-3938.jpg); */
-    /* background-image: url(https://4kwallpapers.com/images/wallpapers/starry-sky-sunlight-sunset-meteorite-fall-5k-1920x1080-158.jpg); */
-    background-image: url(./asset/dusk.jpg);
-    background-size: cover;
-    background-position: center;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-    padding: 2em;
-    width: 100%;
-    height: 100vh;
-    overflow: hidden;
-    font-family: 'Rubik', sans-serif;
-}
+/*
+  https://github.com/tailwindlabs/tailwindcss/discussions/2394
+  https://github.com/tailwindlabs/tailwindcss/pull/5732
+*/
+@layer utilities {
+    /* Chrome, Safari and Opera */
+    .no-scrollbar::-webkit-scrollbar {
+      display: none;
+    }
   
-  .app {
-    background-color: rgba(16, 18, 27, 0.4);
-    /* max-width: 1250px; */
-    /* max-height: 860px; */
-    height: 93vh;
-    width: 90vw;
-    display: flex;
-    flex-direction: column;
-    overflow-x: scroll;
-    overflow-y: hidden;
-    border-radius: 14px;
-    backdrop-filter: blur(10px);
-    -webkit-backdrop-filter: blur(10px);
-    font-size: 15px;
-    font-weight: 500;
-    margin-top: -5%;
-    margin-left: -2.5%;
-    -ms-overflow-style: none; /* for IE/Edge no show scroll bar */
+    .no-scrollbar {
+      -ms-overflow-style: none; /* IE and Edge */
+      scrollbar-width: none; /* Firefox */
+    }
   }
-
-  .app::-webkit-scrollbar {
-    display: none;
-  } /* for Chrome in window os */
-}
-
-
-@media (min-width:480px) and (max-width: 1200px) {
-  body {
-    margin: 0;
-    /* background-image: url(https://4kwallpapers.com/images/wallpapers/starry-sky-sunlight-sunset-meteorite-fall-5k-1920x1080-158.jpg); */
-    background-image: url(./asset/dusk.jpg);
-    background-size: cover;
-    background-position: center;
-    background-repeat: no-repeat;
-    /* display: flex;
-    justify-content: center; */
-    align-items: center;
-    flex-direction: column;
-    /* padding: 2em; */
-    width: 100vw;
-    height: 100vh;
-    min-height: 568px;
-    overflow: hidden;
-  }
-  
-  .app {
-    /* height: 107vh;
-    width: 64vh;
-    margin-top: 0%;
-    margin-left: 0%; */
-
-    background-color: rgba(16, 18, 27, 0.4);
-    height: 95vh;
-    width: 95vw;
-    position: relative;
-    /* left: 3.5%; */
-    display: flex;
-    flex-direction: column;
-    overflow-x: scroll;
-    overflow-y: hidden;
-    border-radius: 14px;
-    backdrop-filter: blur(10px);
-    -webkit-backdrop-filter: blur(10px);
-    font-size: 15px;
-    font-weight: 500;
-    margin-top: 2vh;
-    margin-left: 3vw;
-    -ms-overflow-style: none; /* for IE/Edge no show scroll bar */
-  }
-  .app::-webkit-scrollbar {
-    display: none;
-  } /* for Chrome in window os */
-}
-
-/* @media (max-width: 992px) {
-  .app {
-    height: 105vh;
-    width: auto;
-    margin-top: -10%;
-    margin-left: -5%;
-  }
-}
-
-@media (max-width: 768px) {
-  .app {
-    height: 105vh;
-    width: auto;
-    margin-top: -10%;
-    margin-left: -5%;
-  }
-} */
-
-@media (max-width: 480px) {
-  body {
-    margin: 0;
-    /* background-image: url(https://4kwallpapers.com/images/wallpapers/starry-sky-sunlight-sunset-meteorite-fall-5k-1920x1080-158.jpg); */
-    background-image: url(./asset/dusk.jpg);
-    background-size: cover;
-    background-position: center;
-    background-repeat: no-repeat;
-    /* display: flex;
-    justify-content: center; */
-    align-items: center;
-    flex-direction: column;
-    /* padding: 2em; */
-    width: 100vw;
-    height: 100vh;
-    min-height: 568px;
-    overflow: hidden;
-  }
-  
-  .app {
-    /* height: 107vh;
-    width: 64vh;
-    margin-top: 0%;
-    margin-left: 0%; */
-
-    background-color: rgba(16, 18, 27, 0.4);
-    height: 85vh;
-    width: 90vw;
-    position: relative;
-    /* top: 10vh;
-    left: 3vh; */
-    display: flex;
-    flex-direction: column;
-    overflow-x: scroll;
-    overflow-y: hidden;
-    border-radius: 14px;
-    backdrop-filter: blur(10px);
-    -webkit-backdrop-filter: blur(10px);
-    font-size: 15px;
-    font-weight: 500;
-    margin-top: 5vh;
-    margin-left: 5vw;
-    -ms-overflow-style: none; /* for IE/Edge no show scroll bar */
-  }
-  .app::-webkit-scrollbar {
-    display: none;
-  } /* for Chrome in window os */
-}
